### 2022-12-15 #5020 @editorialbot
[PRE REVIEW]: TopoPyScale: A Python Package for Hillslope Climate Downscaling
**Submitting author:** @arcticsnow (<a href="http://orcid.org/0000-0003-1282-7307">Simon Filhol</a>)
**Repository:** https://github.com/ArcticSnow/TopoPyScale
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
