+++ JOSS/paper.bib
@@ -10,9 +10,9 @@ URL = {https://gmd.copernicus.org/articles/15/1753/2022/},
 DOI = {10.5194/gmd-15-1753-2022}
 }
 
-@ARTICLE{10.3389/feart.2022.789332,
+@ARTICLE{Kruyt2022,
 AUTHOR={Kruyt, Bert and Mott, Rebecca and Fiddes, Joel and Gerber, Franziska and Sharma, Varun and Reynolds, Dylan},   
-TITLE={A Downscaling Intercomparison Study: The Representation of Slope- and Ridge-Scale Processes in Models of Different Complexity},      
+TITLE={A Downscaling Intercomparison Study: The Representation of Slope- and Ridge-Scale Processes in Models of Different Complexity},
 JOURNAL={Frontiers in Earth Science},      
 VOLUME={10},           
 YEAR={2022},      
@@ -106,7 +106,7 @@ volume = {55},
 number = {2},
 pages = {1737-1772},
 keywords = {Earth System Model, hillslope hydrology, Critical Zone, vegetation distribution, global water, energy, carbon cycles, surface water-groundwater interaction},
-doi = {https://doi.org/10.1029/2018WR023903},
+doi = {10.1029/2018WR023903},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2018WR023903},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2018WR023903},
 year = {2019}
@@ -191,7 +191,7 @@ number = {3},
 pages = {123-145},
 year = {2002},
 issn = {0165-232X},
-doi = {https://doi.org/10.1016/S0165-232X(02)00074-5},
+doi = {10.1016/S0165-232X(02)00074-5},
 url = {https://www.sciencedirect.com/science/article/pii/S0165232X02000745},
 author = {Perry Bartelt and Michael Lehning},
 keywords = {Snow, Heat and mass transfer, Snow deformation, Phase change, Finite element model, Avalanche},
