+++ JOSS/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-Global climate reanalyses and projections are available worldwide for the past and coming century. However, model grids remain too coarse to be directly relevant at the hillslope scale [@fanetal2019], requiring adapted downscaling tools to account for the effects of local topography. Mountain regions are experiencing accelerated warming with the cryosphere rapidly responding to climate change. To understand and study geomorphological, hydrological, and glaciological changes, we need tools to downscale meteorological timeseries for the basic atmospheric variables used to the solve surface energy and mass balance in process-based models. Advanced dynamical downscaling methods exist, though they come with a high computational cost and complex technical setup [@10.3389/feart.2022.789332]. `TopoPyScale` uses a pragmatic approach to downscaling by minimizing complexity, reducing computational cost, simplifying interoperability with land surface models, while retaining physical coherence and allowing the primary drivers of land surface-atmosphere interaction to be considered. The toolbox is designed to be flexible in its usage and development. 
+Global climate reanalyses and projections are available worldwide for the past and coming century. However, model grids remain too coarse to be directly relevant at the hillslope scale [@fanetal2019], requiring adapted downscaling tools to account for the effects of local topography. Mountain regions are experiencing accelerated warming with the cryosphere rapidly responding to climate change. To understand and study geomorphological, hydrological, and glaciological changes, we need tools to downscale meteorological timeseries for the basic atmospheric variables used to the solve surface energy and mass balance in process-based models. Advanced dynamical downscaling methods exist, though they come with a high computational cost and complex technical setup [@Kruyt2022]. `TopoPyScale` uses a pragmatic approach to downscaling by minimizing complexity, reducing computational cost, simplifying interoperability with land surface models, while retaining physical coherence and allowing the primary drivers of land surface-atmosphere interaction to be considered. The toolbox is designed to be flexible in its usage and development. 
 
 # Statement of need
 
