### 2022-12-16 #5020 @kthyng
Hi @ArcticSnow and thanks for your submission! It looks to be in pretty good shape, though a couple of quick comments:
1. it'd be better to be able to install your package in one line if possible, through PyPI or conda-forge, for example.
2. examples: I see you have examples but they are annoying to get to — referenced in the docs but then you need to copy and paste a link and then click through several times to get to a notebook. One approach to make this easier is to build the notebooks directly into your docs (readthedocs with notebooks is one option).
3. Can you clarify the difference between the code in this submission and the subject of the 2014 GMD paper?
4. please suggest 5+ reviewers that would be a reasonable fit, you can use this list of volunteers: https://bit.ly/joss-reviewers (but please don't use the "@" so they aren't preemptively pinged).
