### 2022-12-16 #5020 @kthyng
@hugoledoux Is this an ok fit for you? And I remember you saying you're pretty busy at the moment but given the holidays I don't think many people are working a lot at the moment so this assignment could wait until afterward to get started. Would you be able to take it on in the new year?
