### 2022-12-21 #5020 @ArcticSnow
@kthyng @hugoledoux
First of all thank you for taking part in the reviewing process of this manuscript. To answer some of the points listed above:
- (1) I clarified the [install section](https://topopyscale.readthedocs.io/en/latest/1_instal/) of the documentation. I have a technical problem to have a single one liner install has soe libraries such xarray or gdal must be install via conda, and other dependencies are pip only (*e.g.* topocalc). Please let me know if the installation procedure is clear enough as they are now.
- (2) The documentation now includes a step by step on how to run the examples. To be read [here](https://topopyscale.readthedocs.io/en/latest/6_examples/).
- (4) Looking at the people indicating relevant language and expertise domains, I would suggest: rmsare, kmarkert, arbennett, felixcremer, rchg, Mankoff,  fgnievinski, dvalters, sgrieve, Chilipp, andreas-h, josephhardinee, dhhagan, jtmiclat, ashwinvis, kellykochanski

Concerning point (3), we will edit the manuscript as suggested in the coming days.
