### 2022-12-23 #5020 @hugoledoux
this should be sufficient, I assume that the installation will be discussed by the reviewer. For gdal, can't you install [rasterio](https://pypi.org/project/rasterio/) and be done with it?
