### 2022-12-23 0c5dd0e
improved setup.py for one line install. new version
