+++ setup.py
@@ -9,13 +9,18 @@ with open(path.join(this_directory, 'README.md')) as f:
 
 setup(
     name='topopyscale',
-    version='0.1.6',
+    version='0.1.7',
     description='A Python package to perform climate downscaling at the hillslope scale',
     long_description=long_description,
 
     # The project's main homepage.
     url='https://github.com/ArcticSnow/TopoPyScale',
     download_url = 'https://github.com/ArcticSnow/TopoPyScale/releases/latest',
+    project_urls={
+        'Documentation':'https://topopyscale.readthedocs.io/en/latest/',
+        'Source':'https://github.com/ArcticSnow/TopoPyScale',
+        'Examples':'https://github.com/ArcticSnow/TopoPyScale_examples'
+    },
     # Author details
     author=['Simon Filhol', 'Joel Fiddes', 'Kristoffer Aalstad'],
     author_email='simon.filhol@geo.uio.no',
@@ -48,7 +53,21 @@ setup(
     # What does your project relate to?
     keywords=['climate', 'downscaling', 'meteorology', 'xarray'],
     packages=find_packages(),
-    install_requires=['matplotlib', 'scikit-learn', 'pandas', 'numpy', 'netcdf4', 'h5netcdf', 'pvlib', 'topocalc', 'cdsapi',
-                      'rasterio', 'pyproj', 'dask', 'configobj', 'munch'],
+    python_requires='>=3.8',
+    install_requires=['xarray[complete]',
+                        'matplotlib',
+                        'scikit-learn',
+                        'pandas',
+                        'numpy',
+                        'netcdf4',
+                        'h5netcdf',
+                        'pvlib',
+                        'topocalc',
+                        'cdsapi',
+                        'rasterio',
+                        'pyproj',
+                        'dask',
+                        'configobj',
+                        'munch'],
     include_package_data=True
 )
