### 2022-12-23 #5020 @ArcticSnow
I have figured how to technically install xarray fully from pip. The install is pip only given one has a VE ready. Pypi version updated, and documentation too.
