### 2022-12-23 608902b
added how TopoPyScale differs from 2014 version
