+++ JOSS/paper.md
@@ -36,7 +36,7 @@ Global climate reanalyses and projections are available worldwide for the past a
 
 # Statement of need
 
-`TopoPyScale` is a community supported open-source Python package for performing climate downscaling following the initial work of @fiddesTopoSCALEDownscalingGridded2014 and @fiddesTopoSUBToolEfficient2012. It is designed as a toolbox combining computationally efficient methods to downscale climate reanalysis and projections [@hersbachERA5GlobalReanalysis2020]. Atmospheric climate models are run at coarse spatial resolution (25 km for ERA5) missing the heterogeneity imposed by the topography of mountain ranges on atmospheric variables. \autoref{fig:temp_comp} shows an example of a downscaled temperature field for January 1, 2020 at 12:00 computed with `TopoPyScale` using ERA5 data as input. `TopoPyScale` allows one to reconstruct the variability of temperatures observed in between valleys and mountain tops within one single ERA5 gridcell. This method is now used in a number of studies to investigate geophysical processes such as geomorphological dynamics of permafrost [@esurf-2022-39], the hydrology of mountain catchments [@hess-23-4717-2019], mountain glaciers [@tc-2021-380] and downscaling hillslope scale climate projections [@gmd-15-1753-2022]. The ease of use as well as the low computational cost help scientists to quickly obtain hillslope-scale atmospheric forcing data that is representative for their study domain.
+`TopoPyScale` is a community supported open-source Python package for performing climate downscaling following the initial work of @fiddesTopoSCALEDownscalingGridded2014 and @fiddesTopoSUBToolEfficient2012. It is designed as a toolbox combining computationally efficient methods to downscale climate reanalysis and projections [@hersbachERA5GlobalReanalysis2020]. Atmospheric climate models are run at coarse spatial resolution (25 km for ERA5) missing the heterogeneity imposed by the topography of mountain ranges on atmospheric variables. \autoref{fig:temp_comp} shows an example of a downscaled temperature field for January 1, 2020 at 12:00 computed with `TopoPyScale` using ERA5 data as input. `TopoPyScale` allows one to reconstruct the variability of temperatures observed in between valleys and mountain tops within one single ERA5 gridcell. This method is now used in a number of studies to investigate geophysical processes such as geomorphological dynamics of permafrost [@esurf-2022-39], the hydrology of mountain catchments [@hess-23-4717-2019], mountain glaciers [@tc-2021-380] and downscaling hillslope scale climate projections [@gmd-15-1753-2022]. The ease of use as well as the low computational cost help scientists to quickly obtain hillslope-scale atmospheric forcing data that is representative for their study domain. This new implementation of TopoSCALE into `TopoPyScale` brings a new software architecture facilitating development, and usage. It is fully developped within the Python ecosystem of scientific computing libraries, therefore integrating every processing steps into a single workflow.
 
 ![Comparison of the 2 m air temperature on January 1, 2020 at 12:00 UTC over the Southern Carpathians, Romania, in between ERA5 surface product (top) and the downscaled result using TopoPyScale (middle). The DEM (bottom) was segmented using 2000 clusters.\label{fig:temp_comp}](temperature_comparison_crop_scaled.jpg)
 
