### 2022-12-23 #5020 @ArcticSnow
Updated `paper.md` given comments provided above:
- (3) added context as to what is different in between 2014 version and this one in pdf.
