### 2023-01-09 1dfec48
added precipitation lapse rate option, improved plotting functions
