### 2023-01-11 6c1262e
correct bug in get_WMO_observation()
