### 2023-01-24 f3b331e
updated plotting routine, added error print in topo_scale, added function to evaluate optimum number of clusters
