### 2023-01-25 69c3c3c
added function in topoclass for optimum number of clusters search
