### 2023-01-25 a2d16e3
renamed function in topoclass for optimum number of clusters search
