+++ TopoPyScale/topoclass.py
@@ -213,9 +213,9 @@ class Topoclass(object):
 
         '''
         print('---> Testing number of clusters')
-        print(f'Variables used in clustering: {df_param.columns}')
         print(f'Computing scores for {cluster_range} clusters')
         df_param = ts.ds_to_indexed_dataframe(self.toposub.ds_param)
+        print(f'Variables used in clustering: {df_param.columns}')
         df_scaled, self.toposub.scaler = ts.scale_df(df_param)
 
         df_nclusters =ts.find_number_of_clusters(df_scaled,
