### 2023-01-25 2f6e1be
added extracting cluster size stats in search routine
