### 2023-01-25 da26707
clean search number cluster function
