### 2023-02-03 bc20d92
added feature importance to clustering and specify list of clustering features
