+++ TopoPyScale/topo_scale.py
@@ -240,6 +240,7 @@ def downscale_climate(project_directory,
             down_pt['p'] = top.level*(10**2) * np.exp(-(row.elevation-top.z) / (0.5 * (top.t + down_pt.t) * R / g))  # Pressure in bar
 
         # ======= logic  to compute ws, wd without loading data in memory, and maintaining the power of dask
+        down_pt['month'] = ('time', down_pt.time.dt.month.data)
         if precip_lapse_rate_flag:
             monthly_coeffs = xr.Dataset(
                 {
@@ -247,7 +248,6 @@ def downscale_climate(project_directory,
                 },
                 coords={'month': [1,2,3,4,5,6,7,8,9,10,11,12]}
             )
-            down_pt['month'] = ('time', down_pt.time.dt.month.data)
             down_pt['precip_lapse_rate'] = (1 + monthly_coeffs.coef.sel(month=down_pt.month.values).data * (row.elevation - surf_interp.z) * 1e-3) / \
                                            (1 - monthly_coeffs.coef.sel(month=down_pt.month.values).data * (row.elevation - surf_interp.z) * 1e-3)
         else:
@@ -255,8 +255,8 @@ def downscale_climate(project_directory,
 
         down_pt['tp'] = down_pt.precip_lapse_rate * surf_interp.tp  * 1 / tstep_dict.get(tstep) * 10**3 # Convert to mm/hr
         down_pt['theta'] = np.arctan2(-down_pt.u, -down_pt.v)
-        down_pt['theta_neg'] = (down_pt.theta < 0)*(down_pt.theta + 2*np.pi)
-        down_pt['theta_pos'] = (down_pt.theta >= 0)*down_pt.theta
+        down_pt['theta_neg'] = (down_pt.theta < 0) * (down_pt.theta + 2 * np.pi)
+        down_pt['theta_pos'] = (down_pt.theta >= 0) * down_pt.theta
         down_pt = down_pt.drop('theta')
         down_pt['wd'] = (down_pt.theta_pos + down_pt.theta_neg)  # direction in Rad
         down_pt['ws'] = np.sqrt(down_pt.u ** 2 + down_pt.v**2)
