+++ TopoPyScale/topo_sub.py
@@ -40,13 +40,14 @@ def ds_to_indexed_dataframe(ds):
 
 def scale_df(df_param,
              scaler=StandardScaler(),
-             features=None):
+             features={'x':1, 'y':1, 'elevation':4, 'slope':1, 'aspect_cos':1, 'aspect_sin':1, 'svf':1}):
     """
     Function to scale features of a pandas dataframe
 
     Args:
         df_param (dataframe): features to scale
         scaler (scaler object): Default is StandardScaler()
+        features (dict): dictionnary of features to use as predictors with their respect importance. {'x':1, 'y':1}
 
     Returns:
         dataframe: scaled data
@@ -63,13 +64,14 @@ def scale_df(df_param,
 
 def inverse_scale_df(df_scaled,
                      scaler,
-                     features=None):
+                     features={'x':1, 'y':1, 'elevation':4, 'slope':1, 'aspect_cos':1, 'aspect_sin':1, 'svf':1}):
     """
     Function to inverse feature scaling of a pandas dataframe
     
     Args:
         df_scaled (dataframe): scaled data to transform back to original (inverse transfrom)
         scaler (scaler object): original scikit learn scaler
+        features (dict): dictionnary of features to use as predictors with their respect importance. {'x':1, 'y':1}
 
     Returns:
         dataframe: data in original format
@@ -85,7 +87,7 @@ def inverse_scale_df(df_scaled,
 
 
 def kmeans_clustering(df_param,
-                      feature_list=['x','y', 'elevation', 'slope', 'svf', 'aspect_cos', 'aspect_sin'],
+                      features={'x':1, 'y':1, 'elevation':4, 'slope':1, 'aspect_cos':1, 'aspect_sin':1, 'svf':1},
                       n_clusters=100,
                       seed=None,
                       **kwargs):
@@ -94,6 +96,7 @@ def kmeans_clustering(df_param,
 
     Args:
         df_param (dataframe): features
+        features (dict): dictionnary of features to use as predictors with their respect importance. {'x':1, 'y':1}
         n_clusters (int): number of clusters
         seed (int): None or int for random seed generator
         kwargs:
@@ -104,6 +107,7 @@ def kmeans_clustering(df_param,
         dataframe: df_param
 
     """
+    feature_list = features.keys()
     X = df_param[feature_list].to_numpy()
     col_names = df_param[feature_list].columns
     print('---> Clustering with K-means in {} clusters'.format(n_clusters))
@@ -115,13 +119,19 @@ def kmeans_clustering(df_param,
     return df_centers, kmeans, df_param
 
 
-def minibatch_kmeans_clustering(df_param, n_clusters=100, n_cores=4, seed=None, **kwargs):
+def minibatch_kmeans_clustering(df_param,
+                                n_clusters=100,
+                                features={'x':1, 'y':1, 'elevation':4, 'slope':1, 'aspect_cos':1, 'aspect_sin':1, 'svf':1},
+                                n_cores=4,
+                                seed=None,
+                                **kwargs):
     """
     Function to perform mini-batch K-mean clustering
     
     Args:
         df_param (dataframe): features
         n_clusters (int):  number of clusters
+        features (dict): dictionnary of features to use as predictors with their respect importance. {'x':1, 'y':1}
         n_cores (int): number of processor core
         kwargs:
 
@@ -130,7 +140,8 @@ def minibatch_kmeans_clustering(df_param, n_clusters=100, n_cores=4, seed=None, 
         kmean object: kmean model
         dataframe: labels of input data
     """
-    X = df_param.to_numpy()
+    feature_list = features.keys()
+    X = df_param[feature_list].to_numpy()
[…]
