### 2023-02-11 c3022b4
wrote code structure. Start parallelization
