+++ TopoPyScale/topo_scale.py
@@ -68,24 +68,6 @@ def clear_files(path):
             os.unlink(os.path.join(dirpath, filename))
         print(f'{path} cleaned')
 
-def parallelize_downscaling(n_core):
-    '''
-    WARNING: this function is a draft. Implementaiton not finished
-
-    ROADMAP:
-    - switch off parallelize in open_mfdataset in downscale_climate()
-    - remove for loops over point_id in downscale_climate() and insted downscale each point_id indivdually on cores
-    - figure out how to pass function argutmenst. Could build a zip contain
-
-    '''
-    print('WARNING: Feature not finished')
-    fun_param = zip()
-    pool = Pool(n_core)
-    pool.starmap(downscale_climate, fun_param)
-    pool.close()
-    pool.join()
-
-
 
 def downscale_climate(project_directory,
                       df_centroids,
@@ -159,50 +141,37 @@ def downscale_climate(project_directory,
     # ============ Distribute each point on cores ======================================
 
     #------------------------------------ IN CONSTRUCTION
+    # note: have a look at: https://github.com/ArcticSnow/TopoPyScale/blob/tc_pool/TopoPyScale/topo_scale.py
+
     # Replace for loop below by a function
     print('WARNING: Feature not finished')
 
-    def downscale_interp():
-        # insrt here downscaling routine for t,q,u,v,tp,p
-        # save file to tmp/
-
-
-    fun_param = zip()
-    pool = Pool(n_core)
-    pool.starmap(downscale_interp, fun_param)
-    pool.close()
-    pool.join()
-
-    def downscale_radiations():
-        # insrt here downscaling routine for sw and lw
-        # save file final file
-
-    fun_param = zip()
-    pool = Pool(n_core)
-    pool.starmap(downscale_radiations, fun_param)
-    pool.close()
-    pool.join()
-
-
-    ##############################################################################################
-    #       OLD CODE
-    #############################################################################################
-
-    dataset = []
-    dpt_list = []
-    dpt_paths = []
-    surf_paths  = []
-    surf_list = []
-
-    n_digits = len(str(df_centroids.index.max()))
+    # Preparing list to feed into Pooling
+    surf_pt_list = []
+    plev_pt_list = []
+    solar_ds_list = []
+    horizon_da_list = []
+    row_list = []
+    meta_list = []
+    interp_method_list = []
+    lw_terrain_flag_list = []
     for i, row in df_centroids.iterrows():
-        pt_id = np.int(row.point_id)
-        print('Downscaling t,q,u,v,tp,p for point: {} out of {}'.format(pt_id+1, df_centroids.index.max()+1))
+        print('Preparing point {}'.format(row.name))
         # =========== Extract the 3*3 cells centered on a given point ============
         ind_lat = np.abs(ds_surf.latitude-row.y).argmin()
         ind_lon = np.abs(ds_surf.longitude-row.x).argmin()
-        ds_surf_pt = ds_surf.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1])
-        ds_plev_pt = ds_plev.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1])
+        surf_pt_list.append(ds_surf.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]))
[…]
