### 2023-02-14 88a940d
fix export to_netcdf() function when no variables defined
