### 2023-02-14 2c40f94
to_netcdf() was missing default choice of variables
