+++ TopoPyScale/topo_scale.py
@@ -52,8 +52,7 @@ from pyproj import Transformer
 import numpy as np
 import sys, time
 from TopoPyScale import meteo_util as mu
-from multiprocessing.dummy import Pool as ThreadPool
-import multiprocessing as Pool
+from multiprocessing import Pool
 import os
 
 # Physical constants
@@ -80,7 +79,8 @@ def downscale_climate(project_directory,
                       lw_terrain_flag=True,
                       tstep='1H',
                       precip_lapse_rate_flag=True,
-                      file_pattern='down_pt*.nc'):
+                      file_pattern='down_pt*.nc',
+                      n_core=4):
     """
     Function to perform downscaling of climate variables (t,q,u,v,tp,SW,LW) based on Toposcale logic
 
@@ -93,15 +93,21 @@ def downscale_climate(project_directory,
         lw_terrain_flag (bool): flag to compute contribution of surrounding terrain to LW or ignore
         tstep (str): timestep of the input data, default = 1H
         file_pattern (str): filename pattern for storing downscaled points, default = 'down_pt_*.nc'
+        n_core (int): number of core on which to distribute processing
 
     Returns:
         dataset: downscaled data organized with time, point_id, lat, long
     """
+    global pt_downscale_interp
+    global pt_downscale_radiations
+
     print('\n---> Downscaling climate to list of points using TopoScale')
     clear_files(f'{project_directory}outputs/tmp')
 
     start_time = time.time()
     tstep_dict = {'1H': 1, '3H': 3, '6H': 6}
+    n_digits = len(str(df_centroids.index.max()))
+
     # =========== Open dataset with Dask =================
     tvec = pd.date_range(start_date, pd.to_datetime(end_date) + pd.to_timedelta('1D'), freq=tstep, closed='left')
     ds_plev = xr.open_mfdataset(project_directory + 'inputs/climate/PLEV*.nc', parallel=True).sel(time=tvec.values)
@@ -149,32 +155,40 @@ def downscale_climate(project_directory,
     # Preparing list to feed into Pooling
     surf_pt_list = []
     plev_pt_list = []
-    solar_ds_list = []
+    ds_solar_list = []
     horizon_da_list = []
     row_list = []
     meta_list = []
-    interp_method_list = []
-    lw_terrain_flag_list = []
+
+
     for i, row in df_centroids.iterrows():
         print('Preparing point {}'.format(row.name))
         # =========== Extract the 3*3 cells centered on a given point ============
         ind_lat = np.abs(ds_surf.latitude-row.y).argmin()
         ind_lon = np.abs(ds_surf.longitude-row.x).argmin()
-        surf_pt_list.append(ds_surf.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]))
-        plev_pt_list.append(ds_plev.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]))
-        solar_ds_list.append(solar_ds.sel(point_id=row.name))
+        ds_surf.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]).to_netcdf(f'outputs/tmp/ds_surf_pt_{i}.nc')
[…]
