### 2023-02-16 3161b86
parallezed topo_scale data preprocessing
