### 2023-02-16 023b086
set variable to None to improve memory
