+++ TopoPyScale/topo_scale.py
@@ -54,11 +54,44 @@ import sys, time
 from TopoPyScale import meteo_util as mu
 from multiprocessing import Pool
 from multiprocessing.dummy import Pool as ThreadPool
+import multiprocessing as mproc
 import os, glob
 
 # Physical constants
 g = 9.81    #  Acceleration of gravity [ms^-1]
 R = 287.05  #  Gas constant for dry air [JK^-1kg^-1]
+
+def multicore_pooling(fun, fun_param, n_cores):
+    '''
+    Function to perform multiprocessing on n_cores
+    Args:
+        fun (obj): function to distribute
+        fun_param zip(list): zip list of functino arguments
+        n_core (int): number o cores
+    '''
+    if num_cores is None:
+        n_cores = mproc.cpu_count() - 2
+        print(f'WARNING: number of cores to use not provided. By default {n_cores} cores will be used')
+
+    pool = Pool(n_cores)
+    pool.starmap(fun, fun_param)
+    pool.close()
+    pool.join()
+    pool = None
+
+def multithread_pooling(fun, fun_param, n_threads):
+    '''
+    Function to perform multiprocessing on n_threads
+    Args:
+        fun (obj): function to distribute
+        fun_param zip(list): zip list of functino arguments
+        n_core (int): number of threads
+    '''
+    tpool = ThreadPool(n_threads)
+    tpool.starmap(fun, fun_param)
+    tpool.close()
+    tpool.join()
+    tpool = None
 
 def clear_files(path):
     # Clear outputs/tmp/ folder
@@ -151,7 +184,8 @@ def downscale_climate(project_directory,
         # =========== Extract the 3*3 cells centered on a given point ============
         ind_lat = np.abs(ds_.latitude - row.y).argmin()
         ind_lon = np.abs(ds_.longitude - row.x).argmin()
-        ds_.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]).to_netcdf(f'outputs/tmp/ds_{type}_pt_{pt_id}.nc')
+        ds_.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]).to_netcdf(f'outputs/tmp/ds_{type}_pt_{pt_id}.nc', engine='h5netcdf')
+        ds_ = None
 
     ds_plev = _open_dataset_climate(flist_PLEV).sel(time=tvec.values)
 
@@ -162,11 +196,7 @@ def downscale_climate(project_directory,
         ds_list.append(ds_plev)
 
     fun_param = zip(ds_list, row_list,  ['plev']*len(row_list), range(0,len(row_list))) # construct here the tuple that goes into the pooling for arguments
-    tpool = ThreadPool(n_core)
-    tpool.starmap(_subset_climate_dataset, fun_param)
-    tpool.close()
-    tpool.join()
-    tpool = None
+    multithread_pooling(_subset_climate_dataset, fun_param, n_threads=n_core)
     fun_param = None
     ds_plev = None
 
@@ -176,11 +206,7 @@ def downscale_climate(project_directory,
         ds_list.append(ds_surf)
 
     fun_param = zip(ds_list, row_list, ['surf']*len(row_list), range(0,len(row_list))) # construct here the tuple that goes into the pooling for arguments
-    tpool = ThreadPool(n_core)
-    tpool.starmap(_subset_climate_dataset, fun_param)
-    tpool.close()
-    tpool.join()
-    tpool = None
+    multithread_pooling(_subset_climate_dataset, fun_param, n_threads=n_core)
     fun_param = None
[…]
