### 2023-02-18 52a315c
replace os.path with pathlib.Path (Python3 standard); optimize imports
