+++ TopoPyScale/topo_scale.py
@@ -44,29 +44,35 @@ TODO:
     - method (2) concatenate 3*3 grid along an additional dimension to process all points in one go.
 
 """
-import pdb
+import multiprocessing as Pool
+import sys
+import time
+from pathlib import Path
+from typing import Union
 
+import numpy as np
 import pandas as pd
 import xarray as xr
 from pyproj import Transformer
-import numpy as np
-import sys, time
+
 from TopoPyScale import meteo_util as mu
-from multiprocessing.dummy import Pool as ThreadPool
-import multiprocessing as Pool
-import os
 
 # Physical constants
-g = 9.81    #  Acceleration of gravity [ms^-1]
-R = 287.05  #  Gas constant for dry air [JK^-1kg^-1]
+g = 9.81  # Acceleration of gravity [ms^-1]
+R = 287.05  # Gas constant for dry air [JK^-1kg^-1]
 
-def clear_files(path):
+
+def clear_files(path: Union[str, Path]):
+    if isinstance(path, str):
+        path = Path(path)
+
     # Clear outputs/tmp/ folder
-    for dirpath, dirnames, filenames in os.walk(path):
-        # Remove regular files, ignore directories
-        for filename in filenames:
-            os.unlink(os.path.join(dirpath, filename))
+    filelist = [f for f in path.glob('**/*') if f.is_file()]
+
+    for file in filelist:
+        file.unlink()
         print(f'{path} cleaned')
+
 
 def parallelize_downscaling(n_core):
     '''
@@ -84,7 +90,6 @@ def parallelize_downscaling(n_core):
     pool.starmap(downscale_climate, fun_param)
     pool.close()
     pool.join()
-
 
 
 def downscale_climate(project_directory,
@@ -151,7 +156,7 @@ def downscale_climate(project_directory,
 
     # ============ Convert lat lon to projected coordinates ==================
     trans = Transformer.from_crs("epsg:4326", "epsg:" + str(target_EPSG), always_xy=True)
-    nxv,  nyv = np.meshgrid(ds_surf.longitude.values, ds_surf.latitude.values)
+    nxv, nyv = np.meshgrid(ds_surf.longitude.values, ds_surf.latitude.values)
     nlons, nlats = trans.transform(nxv, nyv)
     ds_surf = ds_surf.assign_coords({"latitude": nlats[:, 0], "longitude": nlons[0, :]})
     ds_plev = ds_plev.assign_coords({"latitude": nlats[:, 0], "longitude": nlons[0, :]})
@@ -159,30 +164,31 @@ def downscale_climate(project_directory,
     # ============ Loop over each point ======================================
     # Loop over each points (lat,lon) for which to downscale climate variable using Toposcale method
 
-
     dataset = []
     dpt_list = []
     dpt_paths = []
-    surf_paths  = []
+    surf_paths = []
     surf_list = []
 
     n_digits = len(str(df_centroids.index.max()))
 
     for i, row in df_centroids.iterrows():
         pt_id = np.int(row.point_id)
-        print('Downscaling t,q,u,v,tp,p for point: {} out of {}'.format(pt_id+1, df_centroids.index.max()+1))
+        print('Downscaling t,q,u,v,tp,p for point: {} out of {}'.format(pt_id + 1, df_centroids.index.max() + 1))
         # =========== Extract the 3*3 cells centered on a given point ============
-        ind_lat = np.abs(ds_surf.latitude-row.y).argmin()
-        ind_lon = np.abs(ds_surf.longitude-row.x).argmin()
[…]
