+++ TopoPyScale/topo_obs.py
@@ -3,13 +3,16 @@ Tools to download and compare Downsclaed timeseries to observation
 All observation in folder inputs/obs/
 S. Filhol December 2021
 """
-import requests, os, glob
-import pandas as pd
-import numpy as np
-import xarray as xr
 import warnings
-warnings.filterwarnings("ignore")
+from pathlib import Path
 
+import glob
+import os
+import pandas as pd
+import requests
+import xarray as xr
+
+warnings.filterwarnings("ignore")
 
 
 def get_metno_obs(sources, voi, start_date, end_date, client_id=os.getenv('FROST_API_CLIENTID')):
@@ -34,7 +37,7 @@ def get_metno_obs(sources, voi, start_date, end_date, client_id=os.getenv('FROST
     - convert df to xarray dataset with stn_id, time as coordinates
     """
     endpoint = 'https://frost.met.no/observations/v0.jsonld'
-    
+
     df_out = pd.DataFrame()
     for source in sources:
         print('.............................................')
@@ -47,7 +50,7 @@ def get_metno_obs(sources, voi, start_date, end_date, client_id=os.getenv('FROST
             }
 
             # Issue an HTTP GET request
-            r = requests.get(endpoint, parameters, auth=(client_id,''))
+            r = requests.get(endpoint, parameters, auth=(client_id, ''))
             json = r.json()
             # Check if the request worked, print out any errors
             if r.status_code == 200:
@@ -55,7 +58,7 @@ def get_metno_obs(sources, voi, start_date, end_date, client_id=os.getenv('FROST
                 print('-> got VOI: {}'.format(var))
                 df = pd.DataFrame()
                 for i in range(len(data)):
-                    row = pd.DataFrame(data[i]['observations']) # [:1] # raw data = [:1]; corrected = [1:]
+                    row = pd.DataFrame(data[i]['observations'])  # [:1] # raw data = [:1]; corrected = [1:]
                     row['referenceTime'] = data[i]['referenceTime']
                     row['sourceId'] = data[i]['sourceId']
                     df = df.append(row)
@@ -68,8 +71,9 @@ def get_metno_obs(sources, voi, start_date, end_date, client_id=os.getenv('FROST
                 print('..... Error with variable {}! Returned status code {}'.format(var, r.status_code))
                 print('..... Reason: %s' % json['error']['reason'])
                 print('---> {} for {} skipped'.format(var, source))
-    
+
     return df_out
+
 
 def combine_metno_obs_to_xarray(fnames='metno*.pckl', path='inputs/obs/'):
     """
@@ -83,8 +87,8 @@ def combine_metno_obs_to_xarray(fnames='metno*.pckl', path='inputs/obs/'):
         dataset: dataset will all data organized in timeseries and station number
 
     """
-    df_obs = pd.concat(map(pd.read_pickle, glob.glob(os.path.join('', path + fnames))))
-    df = pd.pivot_table(df_obs, columns=['elementId'],values=['value'], index=['sourceId', 'referenceTime'])
+    df_obs = pd.concat(map(pd.read_pickle, glob.glob(Path(('', path + fnames))))
+    df = pd.pivot_table(df_obs, columns=['elementId'], values=['value'], index=['sourceId', 'referenceTime'])
     ds = df.xs('value', axis=1, level=0).to_xarray()
     return ds
 
@@ -132,17 +136,17 @@ def fetch_WMO_insitu_observations(years, months, bbox, target_path='./inputs/obs
             'year': years,
             'month': months,
             'day': ['01', '02', '03',
-				 '04', '05', '06',
-				 '07', '08', '09',
-				 '10', '11', '12',
[…]
