### 2023-02-18 79351fa
replace os.path with pathlib.Path (Python3 standard); optimize imports
