+++ TopoPyScale/topoclass.py
@@ -11,25 +11,29 @@ project/
     -> output/
 
 """
-import glob, os, sys, shutil
-import pdb
-from munch import DefaultMunch
+import glob
+import os
+import shutil
+import sys
+from pathlib import Path
+from typing import Union
+
+import matplotlib.pyplot as plt
+import numpy as np
 import pandas as pd
 import xarray as xr
-import numpy as np
-import matplotlib.pyplot as plt
-from TopoPyScale import fetch_era5 as fe
-from TopoPyScale import topo_param as tp
-from TopoPyScale import topo_sub as ts
-from TopoPyScale import fetch_dem as fd
-from TopoPyScale import solar_geom as sg
-from TopoPyScale import topo_scale as ta
-from TopoPyScale import topo_export as te
-from TopoPyScale import topo_plot as tpl
-from TopoPyScale import topo_obs as tpo
-
+from munch import DefaultMunch
 from sklearn.preprocessing import StandardScaler
 
+from TopoPyScale import fetch_dem as fd
+from TopoPyScale import fetch_era5 as fe
+from TopoPyScale import solar_geom as sg
+from TopoPyScale import topo_export as te
+from TopoPyScale import topo_obs as tpo
+from TopoPyScale import topo_param as tp
+from TopoPyScale import topo_plot as tpl
+from TopoPyScale import topo_scale as ta
+from TopoPyScale import topo_sub as ts
 
 
 class Topoclass(object):
@@ -37,69 +41,66 @@ class Topoclass(object):
     A python class to bring the typical use-case of toposcale in a user friendly object
     """
 
-    def __init__(self, config_file):
-        
-        try:
+    def __init__(self, config_file: Union[str, Path]):
+        if isinstance(config_file, str):
+            config_file = Path(config_file)
+
+        if config_file.exists():
             with open(config_file, 'r') as f:
                 self.config = DefaultMunch.fromYAML(f)
-        except IOError:
-                print('ERROR: config file does not exist. Check path.')
+        else:
+            raise ValueError('ERROR: config file does not exist. Check path.')
+
+        # define directories
+        project_dir = Path('test')
+        output_dir = Path(project_dir, 'outputs')
+        output_tmp = Path(output_dir, 'tmp')
+        downscaled_dir = Path(output_dir, 'downscaled')
+        input_dir = Path(project_dir, 'inputs')
+        dem_dir = Path(input_dir, 'dem')
+        climate_dir = Path(input_dir, 'climate')
+        climate_tmp = Path(climate_dir, 'tmp')
+
+        # create paths if they do not exist
+        for path in [output_dir, output_tmp, downscaled_dir, input_dir, dem_dir climate_dir, climate_tmp]:
+            if not path.exists():
+                path.mkdir()
+
+        self.config.climate.path = climate_dir
+        self.config.dem.path = dem_dir
 
         if self.config.outputs.file.clean_outputs:
             # remove outputs directory because if results already exist this causes concat of netcdf files
-            try:
-                shutil.rmtree(self.config.project.directory + '/outputs/')
-                print('---> Output directory cleaned')
-            except:
-                os.makedirs('/'.join((self.config.project.directory, 'outputs/')))
+            shutil.rmtree(output_dir)
+            print('---> Output directory cleaned')
 
         # remove output fsm directory
         if self.config.outputs.file.clean_FSM:
             try:
-                shutil.rmtree(self.config.project.directory + '/fsm_sims/')
+                shutil.rmtree(Path(self.config.project.directory, 'fsm_sims'))
                 print('---> FSM directory cleaned')
             except:
                 print("---> no FSM directory to clean")
 
             # remove output fsm directory
             try:
[…]
