### 2023-02-18 f47f1b1
concat paths with pathlib
