### 2023-02-19 85a4149
delete files with same output filename pattern as defined in config (-> output folder does not need to be cleared every time) (line 453-462). & auto format file
