+++ TopoPyScale/topoclass.py
@@ -11,25 +11,29 @@ project/
     -> output/
 
 """
-import glob, os, sys, shutil
-import pdb
-from munch import DefaultMunch
+import glob
+import os
+import re
+import shutil
+import sys
+from pathlib import Path
+
+import matplotlib.pyplot as plt
+import numpy as np
 import pandas as pd
 import xarray as xr
-import numpy as np
-import matplotlib.pyplot as plt
-from TopoPyScale import fetch_era5 as fe
-from TopoPyScale import topo_param as tp
-from TopoPyScale import topo_sub as ts
-from TopoPyScale import fetch_dem as fd
-from TopoPyScale import solar_geom as sg
-from TopoPyScale import topo_scale as ta
-from TopoPyScale import topo_export as te
-from TopoPyScale import topo_plot as tpl
-from TopoPyScale import topo_obs as tpo
-
+from munch import DefaultMunch
 from sklearn.preprocessing import StandardScaler
 
+from TopoPyScale import fetch_dem as fd
+from TopoPyScale import fetch_era5 as fe
+from TopoPyScale import solar_geom as sg
+from TopoPyScale import topo_export as te
+from TopoPyScale import topo_obs as tpo
+from TopoPyScale import topo_param as tp
+from TopoPyScale import topo_plot as tpl
+from TopoPyScale import topo_scale as ta
+from TopoPyScale import topo_sub as ts
 
 
 class Topoclass(object):
@@ -38,12 +42,12 @@ class Topoclass(object):
     """
 
     def __init__(self, config_file):
-        
+
         try:
             with open(config_file, 'r') as f:
                 self.config = DefaultMunch.fromYAML(f)
         except IOError:
-                print('ERROR: config file does not exist. Check path.')
+            print('ERROR: config file does not exist. Check path.')
 
         if self.config.outputs.file.clean_outputs:
             # remove outputs directory because if results already exist this causes concat of netcdf files
@@ -72,7 +76,7 @@ class Topoclass(object):
             os.makedirs('/'.join((self.config.project.directory, 'inputs/')))
         if not os.path.exists('/'.join((self.config.project.directory, 'outputs/'))):
             os.makedirs('/'.join((self.config.project.directory, 'outputs/')))
-        
+
         self.config.climate.path = self.config.project.directory + 'inputs/climate/'
         if not os.path.exists('/'.join((self.config.project.directory, 'inputs/climate/'))):
             os.makedirs('/'.join((self.config.project.directory, 'inputs/climate')))
@@ -89,17 +93,16 @@ class Topoclass(object):
 
         self.config.dem.filepath = self.config.dem.path + self.config.dem.file
         if not os.path.isfile(self.config.dem.filepath):
-            
+
             if self.config.project.extent is not None:
                 self.config.project.extent = dict(zip(['latN', 'latS', 'lonW', 'lonE'], self.config.project.extent))
             else:
                 print('ERROR: no extent provided. Must follow this format: [latN, latS, lonW, lonE]')
                 sys.exit()
 
-            fd.fetch_dem(self.config.dem.path, self.config.project.extent, self.config.dem.epsg, self.config.dem.file)            
+            fd.fetch_dem(self.config.dem.path, self.config.project.extent, self.config.dem.epsg, self.config.dem.file)
         else:
             print('\n---> DEM file found')
-
 
         if self.config.project.extent is not None:
             extent_NSWE = self.config.project.extent
@@ -111,10 +114,11 @@ class Topoclass(object):
         print(self.config.project.extent)
         print('Project lat/lon extent:\n')
         print('\t------------------------------')
[…]
