### 2023-02-22 #5059 @dvalters
Very interesting submission @ArcticSnow! I'm just having a few issues getting the example to run here, I wondered if you could point out where I might be going wrong: https://github.com/ArcticSnow/TopoPyScale/issues/55
