### 2023-02-22 ArcticSnow/TopoPyScale#55 @dvalters
It may be that I'm a bit confused about the config files...

in the TopoPyScale folder, I have a `config.ini` file and in the example datasets folders, (TopoPyScale_examples/ex1_norway etc.) I have a `config_point.yml` and `config_spatial.yml`

Do the contents of the yaml files need to be copied to the example ini file? Or do I need all three config files present for the analysis to work?
