### 2023-02-22 ArcticSnow/TopoPyScale#55 @ArcticSnow
Hi @dvalters ,
1. `config.ini` is deprecated. I will remove it. Thank you for pointing it out.
2. toposcale can be run in two mode. One is downscaling to specific points, and the other is aggregating the DEM with k-means. `ex1_norway` has two config files to show how those two modes are being run.
3. I have never encountered the error you have. I have updated this morning the version of both `topopyscale` and of the example repository (and therefore of their respective config files).

you may try pulling the latest repo or updating the pip install (depending on how you installed it).

In general, config files follow the syntax described in: https://topopyscale.readthedocs.io/en/latest/3_configurationFile/

I hope this helps you and resolve the problem you have, which may come from incompability between the main branch of `topopycale` under development and the example repository
