### 2023-02-22 b6f5f7e
remove deprecated config.ini file
