+++ TopoPyScale/config.ini
@@ -1,37 +1,0 @@
-[main]
-project_name = Retezat
-project_description = Downscaling for the Retezat mountains 2014 - 2015
-project_authors = S. Filhol, J. Fiddes
-project_date = Sept 2021
-project_dir = /home/arcticsnow/Desktop/topoTest/
-
-runmode = basins
-
-start_date = 2014-09-01
-end_date = 2015-08-31
-
-[forcing]
-dataset = era5
-era5_product = reanalysis
-# options for 1, 3, or 6 hrs
-time_step = 6H                                  
-plevels = 300,400,500,600,700,1000
-n_threads_download = 12
-n_cores = 4
-
-dem_file = dem_90m_SRTM_proj_clip2.tif
-dem_epsg = 3844
-
-# increment of the azimuth angle for which to compute the horizon angle 
-horizon_az_inc = 10
-
-[toposcale]
-clustering_method = minibatchkmean
-n_clusters = 50
-interpolation_method = inverse_distance
-pt_list = pt_list.csv
-pt_sampling_method = nearest
-lw_terrain_contribution = true
-
-
-
