### 2023-02-22 ArcticSnow/TopoPyScale#55 @dvalters
Hi @ArcticSnow - I've got further along the processing with the latest pull just now, but am getting an error referring to "reference before assignment" now:

```
---> Downscaling climate to list of points using TopoScale
TopoPyScale_examples/ex1_norway_finse/outputs/tmp cleaned
No ERA5T  PRESSURE data present with additional dimension <expver>
Preparing plev for point 0
Preparing plev for point 1
Preparing plev for point 2
Preparing plev for point 3
Preparing plev for point 4
Preparing plev for point 5
Preparing plev for point 6
Preparing plev for point 7
Preparing plev for point 8
Preparing plev for point 9
No ERA5T  PRESSURE data present with additional dimension <expver>
Preparing surf for point 0
Preparing surf for point 1
Preparing surf for point 2
Preparing surf for point 3
Preparing surf for point 4
Preparing surf for point 5
Preparing surf for point 6
Preparing surf for point 7
Preparing surf for point 8
Preparing surf for point 9
Downscaling t,q,p,tp,ws,wd for point: 1
Downscaling t,q,p,tp,ws,wd for point: 2
Downscaling t,q,p,tp,ws,wd for point: 3
Downscaling t,q,p,tp,ws,wd for point: 4
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
Downscaling t,q,p,tp,ws,wd for point: 5
---> WARNING: Point 9 is -403.0 m lower than the 1000hPa geopotential
=> Values sampled from Psurf and lowest Plevel. No vertical interpolation
Downscaling t,q,p,tp,ws,wd for point: 6
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
Downscaling t,q,p,tp,ws,wd for point: 7
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
Downscaling t,q,p,tp,ws,wd for point: 8
---> WARNING: Point 9 is -11.0 m lower than the 1000hPa geopotential
=> Values sampled from Psurf and lowest Plevel. No vertical interpolation
Downscaling t,q,p,tp,ws,wd for point: 9
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
Downscaling t,q,p,tp,ws,wd for point: 10
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
ERROR: Upper pressure level 600 hPa geopotential is lower than cluster mean elevation
multiprocessing.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File ".miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 51, in starmapstar
    return list(itertools.starmap(args[0], args[1]))
  File "TopoPyScale/TopoPyScale/topo_scale.py", line 302, in pt_downscale_interp
    top = plev_interp.isel(level=ind_z_top)
UnboundLocalError: local variable 'ind_z_top' referenced before assignment
"""

Traceback (most recent call last):
  File "TopoPyScale_examples/ex1_norway_finse/pipeline.py", line 12, in <module>
    mp.downscale_climate()
  File "/TopoPyScale/TopoPyScale/topoclass.py", line 439, in downscale_climate
    ta.downscale_climate(self.config.project.directory,
  File "/TopoPyScale/TopoPyScale/topo_scale.py", line 349, in downscale_climate
    multicore_pooling(pt_downscale_interp, fun_param, n_core)
  File "/dev/TopoPyScale/TopoPyScale/topo_scale.py", line 77, in multicore_pooling
    pool.starmap(fun, fun_param)
  File "/miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 372, in starmap
    return self._map_async(func, iterable, starmapstar, chunksize).get()
  File "/miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 771, in get
    raise self._value
  File "/miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/miniconda3/envs/downscaling/lib/python3.9/multiprocessing/pool.py", line 51, in starmapstar
    return list(itertools.starmap(args[0], args[1]))
  File "/dev/TopoPyScale/TopoPyScale/topo_scale.py", line 302, in pt_downscale_interp
    top = plev_interp.isel(level=ind_z_top)
UnboundLocalError: local variable 'ind_z_top' referenced before assignment
```
