### 2023-02-22 ArcticSnow/TopoPyScale#55 @ArcticSnow
The problem is because the higher pressure level in the climate forcing data is not high enough in elevation in respect to the point it tries downscaling at. This behavior is as expected.

There seem to be a problem either with your config file, or with the input climate dataset, or with the DEM. I am actually surprised your climate forcing for  `ex1_norway` has a 600hPa pressure level when the config file will download climate data up to 700hPa.

https://github.com/ArcticSnow/TopoPyScale_examples/blob/3dee762f8a8918b6ecb585ff340d1bce3ebcf25b/ex1_norway_finse/config_spatial.yml#L25

So the pressure levels must be chosen in relation to the topography. The user must make sure the highest pressure level will be above the topography at all time. There could be a pre-check prior to downscaling (a userfriendly feature to add).
