### 2023-02-23 ArcticSnow/TopoPyScale#55 @dvalters
Thanks for the pointers, I will take a look
