### 2023-02-27 cc99ffc
updating setup.py with version number corresponding to release
