+++ setup.py
@@ -9,7 +9,7 @@ with open(path.join(this_directory, 'README.md')) as f:
 
 setup(
     name='topopyscale',
-    version='0.1.7',
+    version='0.2.0',
     description='A Python package to perform climate downscaling at the hillslope scale',
     long_description=long_description,
 
