+++ TopoPyScale/topo_export.py
@@ -421,9 +421,6 @@ def to_crocus(ds,
                      encoding={'time': {'dtype': 'float64', 'calendar':'standard'},
                      'Tair': {'dtype': 'float64'},
                      'Qair': {'dtype': 'float64'},
-                     'xwind': {'dtype': 'float64'},
-                     'ywind': {'dtype': 'float64'},
-                     'precip': {'dtype': 'float64'},
                      'Snowf': {'dtype': 'float64'},
                      'Rainf': {'dtype': 'float64'},
                      'DIR_SWdown': {'dtype': 'float64'},
