### 2023-03-06 6a4c68e
debuged geopotential computation
