### 2023-03-07 a33e79f
removed bug introduced during storing datasets to_netcdf()
