### 2023-03-07 8ff5478
added logic to also download month of the project endDate
