### 2023-03-08 5aa840d
added to to_netcdf() method to export list of variables
