+++ TopoPyScale/topoclass.py
@@ -625,7 +625,10 @@ class Topoclass(object):
             variables (list str): list of variable to export. Default exports all variables
         """
         if variables is None:
-            variables = list(self.downscaled_pts.keys())
+            if (self.config.outputs.variables.lower() == 'all') or (self.config.outputs.variables is None):
+                variables = list(self.downscaled_pts.keys())
+            else:
+                variables = self.config.outputs.variables
 
         te.to_netcdf(self.downscaled_pts[variables], f'{self.config.project.directory}outputs/' + file_out, variables)
         print('---> File {} saved'.format(file_out))
