### 2023-03-08 f514c3a
update to python 3.9 as xarray does not support 3.8 anymore: https://github.com/pydata/xarray/releases/tag/v2023.02.0
