### 2023-03-09 7f392e9
added an automated max number of cores in case n_cores is more than available
