+++ TopoPyScale/topo_scale.py
@@ -72,6 +72,10 @@ def multicore_pooling(fun, fun_param, n_cores):
     if n_cores is None:
         n_cores = mproc.cpu_count() - 2
         print(f'WARNING: number of cores to use not provided. By default {n_cores} cores will be used')
+    elif n_cores >= mproc.cpu_count():
+        n_cores = mproc.cpu_count() - 2
+        print(f'WARNING: Only {mproc.cpu_count()} core available on this machine, reducing n_cores to {n_cores} ')
+
 
     pool = Pool(n_cores)
     pool.starmap(fun, fun_param)
