### 2023-03-09 #5059 @ArcticSnow
In case you are still testing the library, a key bug (introduced the last month) was identified and corrected this week. I hope this will not hinder your appreciation of the library, if using the latest version (v0.2).
