### 2023-03-09 ef84301
bug in if statement to_netcdf list of vairbles, typo in print statemtement
