### 2023-03-10 13c0b67
headless version of fsm results map function
