### 2023-03-10 a42d0a1
made n_core condition less restrictive for those of us who are core poor :)
