### 2023-03-10 781c1fd
added dtype arg (float32/ int16)
