### 2023-03-10 eed2725
added units and attributes to default downscaled dataset in topo_scale.py
