### 2023-03-13 8a98a75
changed relative path to absolute, added check that start and end date in config is compatible with forcing data
