### 2023-03-14 e41104d
new functionality for realtime applications where current month is re downloaded - this provides update of latest available data for daily runs. Controlled by new keyword: "climate:era5:realtime: True/False" (default: False). There would be smarter and more complex ways to do this (i have a couple ideas) - happy for evolution if anyone interested in this feature.
