### 2023-03-14 2c8e375
Added routine to catch the case that keyword "realtime" doesnt exist in config and set to default "False"
