### 2023-03-15 36da8dd
basin average stats routine on simulation results
