### 2023-03-15 935725b
basin average stats routine on simulation results
