### 2023-03-15 098cd56
basin average stats routine on simulation results
