### 2023-03-16 6f2d203
routines for plotting snow climatology
