+++ TopoPyScale/topo_sim.py
@@ -14,7 +14,10 @@ import rasterio
 from rasterio.enums import Resampling
 from matplotlib import pyplot as plt
 import xarray as xr
-import dask.array as da
+from TopoPyScale import topo_da as da
+import matplotlib.pyplot as plt
+from datetime import datetime
+from matplotlib.backends.backend_pdf import PdfPages
 
 def fsm_nlst(nconfig, metfile, nave):
     """
@@ -524,6 +527,102 @@ def agg_stats(df ):
     weighteddf = df * lp.members
     dfagg = np.sum(weighteddf, 1)/ np.sum(lp.members)
     return (dfagg)
+
+def climatology(ncol):
+    HS = agg_by_var_fsm(ncol)
+    HSdf = agg_stats(HS)
+    
+    # Group the DataFrame by day of the year and calculate the mean
+    HSdf_daily_median = HSdf.groupby(HSdf.index.dayofyear).median()
+    # Group the DataFrame by day of the year and calculate the quantiles
+    HSdf_daily_quantiles = HSdf.groupby(HSdf.index.dayofyear).quantile([0.05, 0.95])
+
+    return HSdf_daily_median, HSdf_daily_quantiles
+
+
+
+def climatology_plot(mytitle, HSdf_daily_median, HSdf_daily_quantiles,  HSdf_realtime=None, plot_show=False):
+
+    with PdfPages("snow_tracker.pdf") as pdf:
+
+        df_shifted = HSdf_daily_median#.shift(-244)
+        df_quantiles_shifted = HSdf_daily_quantiles#.shift(-244)
+
+        # Concatenate the shifted time series such that the last 122 days are first, followed by the first 244 days - water year
+        df_shifted0 = pd.concat([df_shifted[244:], df_shifted[:244]])
+        df_quantiles_shifted1 = pd.concat([df_quantiles_shifted.xs(0.05, level=1)[244:], 
+                                          df_quantiles_shifted.xs(0.05, level=1)[:244]],  axis=0)
+
+        df_quantiles_shifted2 = pd.concat([df_quantiles_shifted.xs(0.95, level=1)[244:], 
+                                          df_quantiles_shifted.xs(0.95, level=1)[:244]], axis=0)
+
+        #s.reset_index(drop=True)   
+        df_shifted0.reset_index(drop=True, inplace=True)   
+
+        # Plot the timeseries and the quantiles
+        plt.figure(figsize=(10, 6))
+        plt.plot(df_shifted0, label='Long-term Median')
+        plt.fill_between(df_shifted0.index, 
+                         df_quantiles_shifted1, 
+                         df_quantiles_shifted2, 
+                         alpha=0.2, label='Long-term 5% - 95% Quantiles')
+
+        # plot realtime data
+        if HSdf_realtime is not None:
+            # get water year
+            thisYear = datetime.now().year
+            thisMonth = datetime.now().month
+
+            wy_start = {9,10,11,12}
+            wy_end = {1, 2, 3, 4, 5,6,7,8}
+
+            if thisMonth in wy_start:
+                thisWaterYear = thisYear
+            else:
+                thisWaterYear = thisYear-1
+
+            plt.plot(HSdf_realtime.values, label = "Water year " + str(thisWaterYear) )
+        
+        # plot fSCA (working on this)
+        #plt.plot(fsca.DOY.astype(int)+(366-244)    , fsca.fSCA, label="fSCA MODIS (0-100%) ") # doy starts at jan 1 for modis data, here it is 1 sept so we shift by 122 days
+
+        plt.xlabel('Day of the year')
[…]
