### 2023-03-16 a57bda7
correct lib import osgeo, added DOY column to modis fSCA output object
