### 2023-03-16 aa27ff1
add fsm_dir as path argument
