### 2023-04-05 #5059 @arbennett
Hi @ArcticSnow - apologies for being so delayed on this, but I was able to finish up my review today. Overall I think this is a nice library that brings tons of value to the table for making climate/reanalysis data more actionable. Thank you for the submission, I could see myself using this in the future! Below is my formal review. I'll open up a few issues on the documentation and examples repositories that correspond to these comments.

## Overview
TopoPyScale is a python package for downscaling meteorological data to fine spatial resolutions. It is built on many of the standard toolkits being used in the geosciences, and implements the core downscaling routines from a previous methodology, TopoScale. TopoPyScale provides a computationally frugal and streamlined workflow, which are both invaluable features for enabling many modeling studies. Additionally the authors provide numerous options for input/output formatting and tweaking the core algorithms.

I found TopoPyScale relatively straightforward to get up and running, but difficult to actually use due to a lack of documentation and incomplete examples. Overall I think this tool would be very valuable for anyone who knows how to use it, so after some changes to the documentation and the examples I would find it suitable for publication in JOSS.

## Major comments

The majority of my comments arose when actually using TopoPyScale on the examples. The only large change I would suggest making to the paper would be a table listing what variables are able to be downscaled, if such a limitation exists.

- Variables in the output are lacking metadata when written as netcdf files.
- The examples have some hard coding of paths and other things, and are lacking documentation in a way that made it difficult for me to get up and running. I would try to possibly streamline this experience and possibly even get a binder link set up so that people just exploring TopoPyScale can try out the examples directly in the browser.
- Example scripts for `norway_finse` runs both the distributed and pointwise downscaling, but they overwrite eachother.
- It wasn't clear to me how to get the full gridded product back out rather than just points. I went looking around in the examples and docs but couldn't readily find information about it. More generally I felt like there were many options in the configuration, but it was hard to know what the effects of changing some of them would be. I see there is some listing of options in the example yaml file in the documentation but I found it hard to scroll around there, and would find it much easier to search documentation where options are explicitly enumerated and described.
- There appear to be no tests, but it seems like with the examples something like a benchmark dataset could be included to make sure that the model operates properly?

## Minor comments
- In `compare_obs.ipynb` in TopoPyScale_examples the matplotlib styles are not installed by default, so I would remove this.
- Also the `metno*.pckl` files for `compare_obs.ipynb` do not seem to exist in the examples.
- On https://topopyscale.readthedocs.io/en/latest/06_examples/, under warning 2:
  "a large amount of verbose" -> "a large amount of verbose output"
