### 2023-04-13 #5059 @ArcticSnow
@arbennett , @dvalters ,
Thank you for your time reviewing and suggesting constructive changes to the paper and the software. I am currently in the field for most of April, and will accomplish all suggestions in May for an updated version.
