### 2023-04-21 309cfb4
xr.core.weighted.DatasetWeighted deprecated for xr.Dataset.weighted. According api docs https://docs.xarray.dev/en/stable/generated/xarray.Dataset.weighted.html#xarray.Dataset.weighted
xr.Dataset.weighted returns xr.core.weighted.DatasetWeighted
