### 2023-04-21 e012aaf
solving the following "coordinate dimension" issue in constructing the xarray dataset, see this issue:
  https://github.com/pydata/xarray/issues/2931 xarray.core.merge.MergeError:

  Error that needed fixing:
  coordinate point_id shares a name with a dataset dimension, but is not a 1D variable along that dimension. This is disallowed by the xarray data model.
