### 2023-05-02 b0060f2
adjust unit compatibility to https://github.com/CryoGrid/CryoGrid.jl/issues/9
