### 2023-05-11 df38328
pd.date_range deprecated arg "closed" changed to "inclusive"
