### 2023-05-11 d9b7440
removed append loop in _open_dataset_climate as this generated memory blowout on large domains. Returned to xr.open_mfdataset see issue: https://github.com/ArcticSnow/TopoPyScale/issues/67#issue-1634262427
other small changes:

1. pd.date_range deprecated arg "closed" changed to "inclusive"

2. deprecated np.int -> np.int32

3. added missing temporal subset l.187
