### 2023-05-11 618968e
xr.open_rasterio deprecated for rioxarray.open_rasterio
requires new packange:
import rioxarray
