### 2023-05-11 00cb84b
added new package requirement
xr.open_rasterio deprecated for rioxarray.open_rasterio

requires new packange:
import rioxarray
