+++ TopoPyScale/topo_sim.py
@@ -14,7 +14,7 @@ import rasterio
 from rasterio.enums import Resampling
 from matplotlib import pyplot as plt
 import xarray as xr
-from TopoPyScale import topo_da as da
+#from TopoPyScale import topo_da as da
 import matplotlib.pyplot as plt
 from datetime import datetime
 from matplotlib.backends.backend_pdf import PdfPages
