+++ environment.yml
@@ -1,0 +1,314 @@
+name: downscaling
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - affine=2.4.0
+  - alsa-lib=1.2.8
+  - arrow-cpp=10.0.1
+  - asttokens=2.0.5
+  - attr=2.5.1
+  - attrs=23.1.0
+  - aws-c-auth=0.6.26
+  - aws-c-cal=0.5.26
+  - aws-c-common=0.8.17
+  - aws-c-compression=0.2.16
+  - aws-c-event-stream=0.2.20
+  - aws-c-http=0.7.7
+  - aws-c-io=0.13.21
+  - aws-c-mqtt=0.8.6
+  - aws-c-s3=0.2.8
+  - aws-c-sdkutils=0.1.9
+  - aws-checksums=0.1.14
+  - aws-crt-cpp=0.19.9
+  - aws-sdk-cpp=1.10.57
+  - backcall=0.2.0
+  - blosc=1.21.3
+  - bokeh=3.1.0
+  - boost-cpp=1.78.0
+  - brotli=1.0.9
+  - brotli-bin=1.0.9
+  - bzip2=1.0.8
+  - c-ares=1.18.1
+  - ca-certificates=2023.5.7
+  - cached-property=1.5.2
+  - cached_property=1.5.2
+  - cairo=1.16.0
+  - cfitsio=4.2.0
+  - cftime=1.6.2
+  - click=8.1.3
+  - click-plugins=1.1.1
+  - cligj=0.7.2
+  - cloudpickle=2.2.1
+  - contourpy=1.0.7
+  - curl=8.0.1
+  - cycler=0.11.0
+  - cytoolz=0.12.0
+  - dask-core=2023.4.1
+  - dbus=1.13.6
+  - decorator=5.1.1
+  - distributed=2023.4.1
+  - executing=0.8.3
+  - expat=2.5.0
+  - flit-core=3.8.0
+  - font-ttf-dejavu-sans-mono=2.37
+  - font-ttf-inconsolata=3.000
+  - font-ttf-source-code-pro=2.038
+  - font-ttf-ubuntu=0.83
+  - fontconfig=2.14.2
+  - fonts-conda-ecosystem=1
+  - fonts-conda-forge=1
+  - fonttools=4.39.3
+  - freetype=2.12.1
+  - freexl=1.0.6
+  - geos=3.11.2
+  - geotiff=1.7.1
+  - gettext=0.21.1
+  - gflags=2.2.2
+  - giflib=5.2.1
+  - glib=2.76.2
+  - glib-tools=2.76.2
+  - glog=0.6.0
+  - graphite2=1.3.13
+  - gst-plugins-base=1.22.0
+  - gstreamer=1.22.0
+  - h5netcdf=1.1.0
+  - h5py=3.8.0
+  - harfbuzz=6.0.0
+  - hdf4=4.2.15
+  - hdf5=1.14.0
+  - icu=72.1
+  - importlib-resources=5.12.0
+  - importlib_metadata=6.6.0
+  - importlib_resources=5.12.0
+  - ipython=8.12.0
+  - jedi=0.18.1
+  - jinja2=3.1.2
+  - joblib=1.2.0
+  - json-c=0.16
+  - kealib=1.5.0
+  - keyutils=1.6.1
+  - kiwisolver=1.4.4
+  - krb5=1.20.1
+  - lame=3.100
+  - lcms2=2.15
+  - ld_impl_linux-64=2.38
+  - lerc=4.0.0
+  - libabseil=20230125.0
+  - libaec=1.0.6
+  - libarrow=10.0.1
+  - libblas=3.9.0
[…]
