### 2023-05-11 a739924
edited environment.yml. move some libraries to conda install, remove prefix, and all aws- libraries
