### 2023-05-11 c46d536
Update test_topopyscale.yml
test if env exist
