+++ .github/workflows/test_topopyscale.yml
@@ -20,7 +20,7 @@ jobs:
         echo $CONDA/bin >> $GITHUB_PATH
     - name: Install dependencies
       run: |
-        pip install mamba
+        conda install mamba -n base -c conda-forge
         if conda info --envs | grep -q downscaling; then echo "downscaling already exists" | conda env update --file environment.yml; else mamba env create -f environment.yml; fi
         conda activate downscaling
     - name: Install TopoPyscale main development branch
