+++ .github/workflows/test_topopyscale.yml
@@ -21,9 +21,9 @@ jobs:
     - name: Install dependencies
       run: |
         conda install mamba -n base -c conda-forge
-        if conda info --envs | grep -q downscaling; then echo "downscaling already exists" | conda env update --file environment.yml; else mamba env create -f environment.yml; fi
-        conda init bash
-        conda activate downscaling
+        # if conda info --envs | grep -q downscaling; then echo "downscaling already exists" | conda env update --file environment.yml; else mamba env create -f environment.yml; fi
+        mamba env update --file environment.yml
+        # conda activate downscaling
     - name: Install TopoPyscale main development branch
       run: |
         pip install -e .
