### 2023-05-11 f22fcbe
Update test_topopyscale.yml
added shell
