+++ .github/workflows/test_topopyscale.yml
@@ -5,6 +5,9 @@ on: [push]
 jobs:
   build-linux:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        shell: bash -el {0}
     strategy:
       max-parallel: 5
 
