+++ .github/workflows/test_topopyscale.yml
@@ -4,6 +4,7 @@ on: [push]
 
 jobs:
   build-linux:
+    if: "contains(github.event.commits[0].message, '[trigger test]')"
     runs-on: ubuntu-latest
     defaults:
       run:
@@ -26,10 +27,10 @@ jobs:
         conda install mamba -n base -c conda-forge
         # if conda info --envs | grep -q downscaling; then echo "downscaling already exists" | conda env update --file environment.yml; else mamba env create -f environment.yml; fi
         mamba env update --file environment.yml
-        # conda activate downscaling
     - name: Install TopoPyscale main development branch
       run: |
-        conda activate downscaling
+        conda info
+        $CONDA/bin/activate downscaling
         pip install -e .
     - name: Lint with flake8
       run: |
