+++ .github/workflows/test_topopyscale.yml
@@ -25,12 +25,9 @@ jobs:
     - name: Install dependencies
       run: |
         conda install mamba -n base -c conda-forge
-        # if conda info --envs | grep -q downscaling; then echo "downscaling already exists" | conda env update --file environment.yml; else mamba env create -f environment.yml; fi
-        mamba env update --file environment.yml
+        mamba env update --file environment.yml --name base
     - name: Install TopoPyscale main development branch
       run: |
-        conda info
-        $CONDA/bin/activate downscaling
         pip install -e .
     - name: Lint with flake8
       run: |
