### 2023-05-11 312e085
corrected undefined variables detected by flake8 package. Two issue open: #77 and #76. [trigger test]
