+++ .github/workflows/test_topopyscale.yml
@@ -36,11 +36,6 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    # pytest tests not yet implemented. Need to implement a test script: https://www.youtube.com/watch?v=DhUpxWjOhME
-    #- name: Test with pytest
-      #run: |
-        #conda install pytest
-        #pytest
     - name: Run TopoPyScale example
       run: |
         git clone https://github.com/ArcticSnow/TopoPyScale_examples.git
