### 2023-05-12 99979c2
updated agg_by_var_fsm functions: corrected back column index, added functionality to use variable name rather than only implicit column index
