### 2023-05-12 18ff80a
added functionality to read FSM output as xarray Dataset
