+++ TopoPyScale/topo_sim.py
@@ -5,7 +5,7 @@ J. Fiddes, February 2022
 TODO:
 
 """
-import os
+import os, re
 import glob
 import re
 import pandas as pd
@@ -18,7 +18,7 @@ import xarray as xr
 import matplotlib.pyplot as plt
 from datetime import datetime
 from matplotlib.backends.backend_pdf import PdfPages
-
+import datetime as dt
 def fsm_nlst(nconfig, metfile, nave):
     """
     Function to generate namelist parameter file that is required to run the FSM model.
@@ -81,6 +81,82 @@ def fsm_nlst(nconfig, metfile, nave):
     f.close()
 
 
+def txt2ds(fname):
+    '''
+    Function to read a single FSM text file output as a xarray dataset
+    Args:
+        fname (str): filename
+
+    Returns:
+        xarray dataset of dimension (time, point_id)
+    '''
+    df = read_pt_fsm(fname)
+    point_id = int( re.findall(r'\d+', fname.split('/')[-1])[-1])
+    print(f'---> Reading FSM data for point_id = {point_id}')
+
+    ds = xr.Dataset({
+        "albedo": (['time'], df.albedo.values),
+        "runoff":  (['time'], df.runoff.values),
+        "snd":  (['time'], df.snd.values),
+        "swe":  (['time'], df.swe.values),
+        "t_surface":  (['time'], df.t_surface.values),
+        "t_soil":  (['time'], df.t_soil.values),
+        },
+        coords={
+            "point_id": point_id,
+            "time": df.index,
+            "reference_time": pd.Timestamp(df.index[0])
+        })
+
+    return ds
+
+def to_dataset(fname_pattern='sim_FSM_pt*.txt', fsm_path = "./fsm_sims/"):
+    '''
+    Function to read FSM outputs of one simulation into a single dataset.
+    Args:
+        fname_pattern:
+        fsm_path:
+
+    Returns:
+
+    '''
+    fnames = glob.glob(fsm_path + fname_pattern)
+    fnames.sort()
+
+    ds = xr.concat([txt2ds(fname) for fname in fnames],'point_id')
+
+    ds.albedo.attrs = {'units':'ratio', 'standard_name':'albedo', 'long_name':'Surface Effective Albedo', '_FillValue': -9999999.0}
+    ds.runoff.attrs = {'units':'kg m-2', 'standard_name':'runoff', 'long_name':'Cumulated runoff from snow', '_FillValue': -9999999.0}
+    ds.snd.attrs = {'units':'m', 'standard_name':'snd', 'long_name':'Average snow depth', '_FillValue': -9999999.0}
+    ds.swe.attrs = {'units':'kg m-2', 'standard_name':'swe', 'long_name':'Average snow water equivalent', '_FillValue': -9999999.0}
+    ds.t_surface.attrs = {'units':'°C', 'standard_name':'t_surface', 'long_name':'Average surface temperature', '_FillValue': -9999999.0}
+    ds.t_surface.attrs = {'units':'°C', 'standard_name':'t_soil', 'long_name':'Average soil temperature at 20 cm depth', '_FillValue': -9999999.0}
+    ds.attrs = {'title':'FSM simulation outputs',
+                'source': 'Data downscaled with TopoPyScale and simulated with FSM',
+                'date_created':dt.datetime.now().strftime('%Y/%m/%d %H:%M:%S')}
+
+    return ds
+
+
+def read_pt_fsm(fname):
+    '''
+    Function to load FSM simulation output into a pandas dataframe
+    Args:
+        fname (str): path to simulation file to open
+
+    Returns:
[…]
