### 2023-05-12 cd15985
generalize map_variable() function
