+++ TopoPyScale/topo_plot.py
@@ -7,6 +7,7 @@ S. Filhol, December 2021
 import matplotlib.pyplot as plt
 from matplotlib.colors import LightSource
 import numpy as np
+import xarray as xr
 
 def map_terrain(ds_param,
                 var='elevation',
@@ -48,7 +49,7 @@ def map_terrain(ds_param,
 
 def map_variable(ds_down,
                  ds_param,
-                 time_step=1,
+                 time_step=None,
                  time=None,
                  var='t',
                  ax=None,
@@ -59,11 +60,11 @@ def map_variable(ds_down,
     Function to plot unclustered downscaled points given that each point corresponds to a cluster label
 
     Args:
-        ds_down (dataset): TopoPyScale downscaled point object with coordinates (time, point_id)
+        ds_down (dataset): TopoPyScale downscaled point object with coordinates (time, point_id). Can be single variable or multiple
         ds_param (dataset): TopoPyScale toposub dataset with coordinates (x,y)
         time_step (int): (optional) time step to plot.
         time (str): (optional) time slice to plot. time overule time_step
-        var (str): variable from ds_down to plot. e.g. 't' for temperature
+        var (str): variable from ds_down to plot. e.g. 't' for temperature if ds_down has more than one variable
         cmap (obj): pyplot colormap object
         hillshade (bool): add terrain hillshade in background
         **kwargs: kwargs to pass to imshow() in dict format.
@@ -74,8 +75,14 @@ def map_variable(ds_down,
     if ax is None:
         fig, ax = plt.subplots(1,1)
 
-    if time is None:
-        time = ds_down.time.isel(time=time_step).data
+    if time is None and time_step is not None:
+        ds = ds_down.isel(time=time_step)
+    elif time is not None and time_step is None:
+        ds = ds_down.sel(time=time)
+    elif type(ds_down) is xr.DataArray:
+        ds = ds_down.to_dataset()
+    else:
+        ds = ds_down
 
     if var=='t':
         print('Suggestion: use divergent colormap plt.cm.RdBl_r centered around 0C')
@@ -92,7 +99,10 @@ def map_variable(ds_down,
                    cmap=plt.cm.gray)
         alpha=0.5
 
-    ds_down[var].sel(point_id=ds_param.cluster_labels).sel(time=time).plot.imshow(alpha=alpha, cmap=cmap, **kwargs)
+    if len(list(ds.keys()))==1:
+        var = list(ds.keys())[0]
+    ds[var].sel(point_id=ds_param.cluster_labels).plot.imshow(alpha=alpha, cmap=cmap, **kwargs)
+
 
     return ax
 
