### 2023-05-12 7518a03
handle case where project directory is not given
