### 2023-05-12 424dd88
added logic to make sure n_cores is at least 1 [trigger test]
