### 2023-05-13 #5059 @ArcticSnow
# Revision in Progress

## Minor comments:
- [x] *In compare_obs.ipynb in TopoPyScale_examples the matplotlib styles are not installed by default, so I would remove this.* -> line removed
- [x] *Also the metno*.pckl files for compare_obs.ipynb do not seem to exist in the examples.* -> added to the github repo [here](https://github.com/ArcticSnow/TopoPyScale_examples/tree/main/ex1_norway_finse/inputs/obs)
- [x] *On https://topopyscale.readthedocs.io/en/latest/06_examples/, under warning 2:
    "a large amount of verbose" -> "a large amount of verbose output"* -> Corrected

## Major comments:
- [x] *Example scripts for norway_finse runs both the distributed and pointwise downscaling, but they overwrite eachother*. -> now those are split into two scripts `pipeline_spatial.py` and `pipeline_point.py`. `pipeline_testing.py` was also added to run a test in [github Action](https://github.com/ArcticSnow/TopoPyScale/blob/main/.github/workflows/test_topopyscale.yml).
- [x] *There appear to be no tests, but it seems like with the examples something like a benchmark dataset could be included to make sure that the model operates properly?* -> Github Action script [test_topopyscale.yml](https://github.com/ArcticSnow/TopoPyScale/blob/main/.github/workflows/test_topopyscale.yml) setup a VE from `environment.yml` file, lint the package with Flake8, clone the latest `ex1_norway_finse` example, and finally run the script [`pipeline_test.py`](https://github.com/ArcticSnow/TopoPyScale_examples/blob/main/ex1_norway_finse/pipeline_test.py). This is a basic spatialized downscaling script over two months.
- [x]  *It wasn't clear to me how to get the full gridded product back out rather than just points. I went looking around in the examples and docs but couldn't readily find information about it.* -> A new part is now added to the [plotting](https://topopyscale.readthedocs.io/en/latest/02_quickstart/#plotting) section.
- [x] *More generally I felt like there were many options in the configuration, but it was hard to know what the effects of changing some of them would be. I see there is some listing of options in the example yaml file in the documentation but I found it hard to scroll around there, and would find it much easier to search documentation where options are explicitly enumerated and described.* -> A whole new [table](https://topopyscale.readthedocs.io/en/latest/03_configurationFile/#possible-values-of-configurations) was added to explicit all options available for the `config.yml` file.
- [ ] Variables in the output are lacking metadata when written as netcdf files.
- [ ] The examples have some hard coding of paths and other things, and are lacking documentation in a way that made it difficult for me to get up and running. I would try to possibly streamline this experience and possibly even get a binder link set up so that people just exploring TopoPyScale can try out the examples directly in the browser.
- [ ] The only large change I would suggest making to the paper would be a table listing what variables are able to be downscaled, if such a limitation exists.
