+++ TopoPyScale/topo_plot.py
@@ -51,7 +51,7 @@ def map_variable(ds_down,
                  ds_param,
                  time_step=None,
                  time=None,
-                 var='t',
+                 var=None,
                  ax=None,
                  cmap=plt.cm.RdBu_r,
                  hillshade=True,
