### 2023-05-16 d325946
convert to open_mfdataset()
