### 2023-05-16 4fad2bb
added table to paper [trigger pdf]
