### 2023-05-22 #5059 @ArcticSnow
Dear all,
You will find an updated and revised version of the code (V0.2.2) and the paper following all the constructive suggestions. I hope this fulfills the requirements and is satisfactory for future users of the library.   You may see the checklist above in this discussion thread. I'd be happy to know if you identify further obvious improvements to perform.
Thank you
