### 2023-05-22 77fd990
possibility to provide alternative paths to store climate data, downscaled data and points-csv
