+++ TopoPyScale/topoclass.py
@@ -51,13 +51,15 @@ class Topoclass(object):
         except IOError:
             print(f'ERROR: config file does not exist. \n\t Current file path: {config_file}\n\t Current working directory: {os.getcwd()}')
 
+        output_dir = Path(self.config.project.directory, 'outputs')
+
         if self.config.outputs.file.clean_outputs:
             # remove outputs directory because if results already exist this causes concat of netcdf files
             try:
-                shutil.rmtree(self.config.project.directory + '/outputs/')
+                shutil.rmtree(output_dir)
                 print('---> Output directory cleaned')
             except:
-                os.makedirs('/'.join((self.config.project.directory, 'outputs/')))
+                os.makedirs(output_dir)
 
         # remove output fsm directory
         if self.config.outputs.file.clean_FSM:
@@ -73,25 +75,29 @@ class Topoclass(object):
                 print("---> Ensemble directory cleaned")
             except:
                 print("---> no ensemble directory to clean")
+
+        # directory where to store the downscaled data
+        if self.config.outputs.directory:
+            self.config.outputs.downscaled = Path(self.config.outputs.directory)
+        else:
+            self.config.outputs.downscaled = Path(output_dir, 'downscaled')
+        os.makedirs(self.config.outputs.downscaled, exist_ok=True)
+
+        # climate path
+        if os.path.isabs(self.config.climate.era5.path):
+            self.config.climate.path = self.config.climate.era5.path
+        else:
+            self.config.climate.path = '/'.join((self.config.project.directory, 'inputs/climate/'))
+
         # check if tree directory exists. If not create it
-        if not os.path.exists('/'.join((self.config.project.directory, 'inputs/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'inputs/')))
-        if not os.path.exists('/'.join((self.config.project.directory, 'outputs/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'outputs/')))
+        os.makedirs(self.config.climate.path, exist_ok=True)
+        os.makedirs('/'.join((self.config.project.directory, 'inputs/climate/tmp')), exist_ok=True)
+        os.makedirs('/'.join((self.config.project.directory, 'outputs/tmp')), exist_ok=True)
+        os.makedirs('/'.join((self.config.project.directory, 'outputs/downscaled')), exist_ok=True)
 
-        self.config.climate.path = self.config.project.directory + 'inputs/climate/'
-        if not os.path.exists('/'.join((self.config.project.directory, 'inputs/climate/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'inputs/climate')))
-        if not os.path.exists('/'.join((self.config.project.directory, 'inputs/climate/tmp/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'inputs/climate/tmp')))
-        if not os.path.exists('/'.join((self.config.project.directory, 'outputs/tmp/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'outputs/tmp')))
-        if not os.path.exists('/'.join((self.config.project.directory, 'outputs/downscaled'))):
-            os.makedirs('/'.join((self.config.project.directory, 'outputs/downscaled')))
-
-        self.config.dem.path = self.config.project.directory + '/inputs/dem/'
-        if not os.path.exists('/'.join((self.config.project.directory, 'inputs/dem/'))):
-            os.makedirs('/'.join((self.config.project.directory, 'inputs/dem')))
+        if not self.config.dem.path:
+            self.config.dem.path = self.config.project.directory + '/inputs/dem/'
+        os.makedirs(self.config.dem.path, exist_ok=True)
 
         self.config.dem.filepath = self.config.dem.path + self.config.dem.file
[…]
