### 2023-05-23 d74bd37
committed this issue https://github.com/ArcticSnow/TopoPyScale/issues/40
