### 2023-05-23 5e1aed2
updated climate path arg to default to original default inputs/climate so not to break existing configs
