### 2023-05-23 #5059 @arbennett
Awesome, thank you so much @hugoledoux and @ArcticSnow! The changes look great, I'm particularly thankful for the table of configuration options in the docs and the updated example notebooks. I'm happy to say that I think this is ready to be accepted!
