### 2023-05-24 5dc0f57
option do name downscaled points by csv column
