+++ TopoPyScale/topo_scale.py
@@ -47,10 +47,13 @@ from multiprocessing import Pool
 from multiprocessing.dummy import Pool as ThreadPool
 import multiprocessing as mproc
 import os, glob
+from pathlib import Path
+from typing import Union
 
 # Physical constants
-g = 9.81    #  Acceleration of gravity [ms^-1]
-R = 287.05  #  Gas constant for dry air [JK^-1kg^-1]
+g = 9.81  # Acceleration of gravity [ms^-1]
+R = 287.05  # Gas constant for dry air [JK^-1kg^-1]
+
 
 def multicore_pooling(fun, fun_param, n_cores):
     '''
@@ -71,12 +74,12 @@ def multicore_pooling(fun, fun_param, n_cores):
     if n_cores == 0:
         n_cores = 1
 
-
     pool = Pool(n_cores)
     pool.starmap(fun, fun_param)
     pool.close()
     pool.join()
     pool = None
+
 
 def multithread_pooling(fun, fun_param, n_threads):
     '''
@@ -92,13 +95,15 @@ def multithread_pooling(fun, fun_param, n_threads):
     tpool.join()
     tpool = None
 
-def clear_files(path):
-    # Clear outputs/tmp/ folder
-    for dirpath, dirnames, filenames in os.walk(path):
-        # Remove regular files, ignore directories
-        for filename in filenames:
-            os.unlink(os.path.join(dirpath, filename))
-        print(f'{path} cleaned')
+
+def clear_files(path: Union[Path, str]):
+    if not isinstance(path, Path):
+        path = Path(path)
+
+    files = [f for f in path.rglob('*') if f.is_file()]
+    for file in files:
+        file.unlink()
+    print(f'{path} cleaned')
 
 
 def downscale_climate(project_directory,
@@ -150,13 +155,9 @@ def downscale_climate(project_directory,
     flist_PLEV = glob.glob(f'{climate_directory}/PLEV*.nc')
     flist_SURF = glob.glob(f'{climate_directory}/SURF*.nc')
 
-
     def _open_dataset_climate(flist):
-    
 
         ds_ = xr.open_mfdataset(flist, parallel=True)
-
-
 
         # this block handles the expver dimension that is in downloaded ERA5 data if data is ERA5/ERA5T mix. If only ERA5 or
         # only ERA5T it is not present. ERA5T data can be present in the timewindow T-5days to T -3months, where T is today.
@@ -179,15 +180,17 @@ def downscale_climate(project_directory,
         # =========== Extract the 3*3 cells centered on a given point ============
         ind_lat = np.abs(ds_.latitude - row.lat).argmin()
         ind_lon = np.abs(ds_.longitude - row.lon).argmin()
-        #from remote_pdb import set_trace
-        #set_trace()
-        ds_tmp = ds_.isel(latitude=[ind_lat-1, ind_lat, ind_lat+1], longitude=[ind_lon-1, ind_lon, ind_lon+1]).copy()
+        # from remote_pdb import set_trace
+        # set_trace()
+        ds_tmp = ds_.isel(latitude=[ind_lat - 1, ind_lat, ind_lat + 1],
+                          longitude=[ind_lon - 1, ind_lon, ind_lon + 1]).copy()
         # convert geopotential height to elevation (in m), normalizing by g
-        ds_tmp['z'] = ds_tmp.z/g
+        ds_tmp['z'] = ds_tmp.z / g
 
         comp = dict(zlib=True, complevel=5)
         encoding = {var: comp for var in ds_tmp.data_vars}
[…]
