### 2023-05-26 11b7b01
new function to export grids of simulation results (based on gridded forcing method). Update on netcdf output method to make it more flexible with respect to single or multiple variables.
