### 2023-06-01 9779e63
point_id always as string
