### 2023-06-01 70d5249
point_id as string (and float to int to str)
