+++ TopoPyScale/topoclass.py
@@ -319,7 +319,8 @@ class Topoclass(object):
             tmp = df_param.groupby('cluster_labels').mean()
             for var in df_param.drop(flist, axis=1).columns:
                 self.toposub.df_centroids[var] = tmp[var]
-        self.toposub.df_centroids['point_id'] = self.toposub.df_centroids.index.astype(int)
+        n_digits = len(str(self.toposub.df_centroids.index.max()))
+        self.toposub.df_centroids['point_id'] = self.toposub.df_centroids.index.astype(int).str.zfill(n_digits)
         self.toposub.ds_param['cluster_labels'] = (
             ["y", "x"], np.reshape(df_param.cluster_labels.values, self.toposub.ds_param.slope.shape))
 
