### 2023-06-01 ea0dc32
option do name downscaled points by csv column
