### 2023-06-01 d858702
point ID in netcdf as integer (if its a number)
