### 2023-06-01 e86a751
changed default behavious iof opening ncdf to False from True. True is giving random errors as refernce here https://github.com/ArcticSnow/TopoPyScale/issues/80. If True is needed and works OK in other circumstances we should consider a config item
