### 2023-06-01 241a37f
added change resolution option to fetch_dem routine
