### 2023-06-01 c235733
added change resolution option to fetch_dem routine
