### 2023-06-01 e4da4a1
add epsg attribute to write_netcdf function that outputs grids
