+++ TopoPyScale/topoclass.py
@@ -85,7 +85,8 @@ class Topoclass(object):
             self.config.outputs.downscaled = Path(self.config.outputs.directory)
         else:
             self.config.outputs.downscaled = self.config.outputs.path / 'downscaled'
-
+        
+        print(self.config.outputs.downscaled)
         # climate path
         if os.path.isabs(self.config.climate.era5.path):
             self.config.climate.path = self.config.climate.era5.path
