### 2023-06-02 420f80a
additional atrributes for output netcdf gridded results datasets
