### 2023-06-02 #5059 @ArcticSnow
Thank you @arbennett and @dvalters  for your positive responses to the latest changes. @hugoledoux, do I need to do something to pursue further the publishing process in JOSS?
