### 2023-06-02 0d2b632
added product type option for WMO observation download
