+++ TopoPyScale/topo_da.py
@@ -14,6 +14,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 import copy
 import rasterio as rio
+from rasterio.plot import show
 from osgeo import gdal
 from osgeo import osr 
 from scipy.interpolate import interp1d
@@ -21,6 +22,7 @@ from TopoPyScale import topo_sim as sim
 from datetime import datetime
 import xarray as xr
 import rioxarray
+from scipy.special import logsumexp
 
 
 # import gdal
@@ -295,6 +297,7 @@ def get_modis_wrapper(vertmax, horizmax, vertmin, horizmin, startDate, endDate):
 
 def pymodis_download(wdir, vert, horiz, STARTDATE, ENDDATE):
     # http://www.pymodis.org/scripts/modis_download.html
+    # pip inpip install pyModis
 
     USER = "jfiddes"  # os.getenv('user')
     PWD = "sT0kkang!"  # os.getenv('pwd')
@@ -577,9 +580,7 @@ def particle_batch_smoother(Y, HX, R):
     w = LH / sum(LH)
 
 
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.special import logsumexp
+
 
 
 def EnKA(prior, obs, pred, alpha, R):
@@ -737,7 +738,7 @@ def PBS(obs, pred, R):
         residual = obs - pred
         llh = -0.5 * ((residual ** 2) * (1 / R))
     else:
-        residual = obs[:, None] - pred
+        residual = np.array(obs)[:, None] - pred
         # pdb.set_trace()
         llh = -0.5 * ((1 / R) @ (residual ** 2))
 
@@ -867,7 +868,7 @@ def fsca_plots(wdir, plotDay, df_mean):
     src = rio.open(modis_file[0])
     fig, (axrgb, axhist) = plt.subplots(1, 2, figsize=(14,7))
 
-    rio.plot.show(src, ax=axrgb,vmin=0, vmax=100)
+    show(src, ax=axrgb,vmin=0, vmax=100)
     cbar1 = plt.imshow(src.read(1), cmap='viridis', vmin=0, vmax=100)
     fig.colorbar(cbar1, ax=axrgb)
 
@@ -889,7 +890,7 @@ def fsca_plots(wdir, plotDay, df_mean):
         array[array > 0] = 1
 
 
-    # rasterio.plot.show(array, cmap='viridis')
+    # rasteshow(array, cmap='viridis')
     # plt.show()
 
     with rio.open('output_raster.tif', 'w', **profile) as dst:
@@ -897,7 +898,7 @@ def fsca_plots(wdir, plotDay, df_mean):
         dst.write(array.astype(rio.int16))
 
     src1 = rio.open("output_raster.tif")
-    rio.plot.show(src1, ax=axhist)
+    show(src1, ax=axhist)
     cbar2 = plt.imshow(src1.read(1), cmap='viridis')
     fig.colorbar(cbar2, ax=axhist)
     plt.show()
@@ -944,7 +945,7 @@ def fsca_plots(wdir, plotDay, df_mean):
 #         array[array > 0] = 1
 #
 #
-#     # rasterio.plot.show(array, cmap='viridis')
+#     # rasteshow(array, cmap='viridis')
 #     # plt.show()
 #
 #     with rio.open('output_raster.tif', 'w', **profile) as dst:
@@ -953,7 +954,7 @@ def fsca_plots(wdir, plotDay, df_mean):
 #
 #     if plot:
 #         src1 = rio.open("output_raster.tif")
-#         rio.plot.show(src1, ax=axhist)
+#         show(src1, ax=axhist)
 #         cbar2 = plt.imshow(src1.read(1), cmap='viridis')
[…]
