### 2023-06-07 7da3632
Added attributes to grid data netcdf export routine
