+++ TopoPyScale/topo_scale.py
@@ -440,7 +440,7 @@ def downscale_climate(project_directory,
             down_pt['LW'] = row.svf * surf_interp['aef'] * sbc * down_pt.t ** 4
 
         kt = surf_interp.ssrd * 0
-        sunset = ds_solar.sunset.astype(bool)
+        sunset = ds_solar.sunset.astype(bool).compute()
         mu0 = ds_solar.mu0
         SWtoa = ds_solar.SWtoa
 
@@ -466,7 +466,9 @@ def downscale_climate(project_directory,
         down_pt['cos_illumination'] = down_pt.cos_illumination_tmp * (
                 down_pt.cos_illumination_tmp > 0)  # remove selfdowing ccuring when |Solar.azi - aspect| > 90
         down_pt = down_pt.drop(['cos_illumination_tmp'])
-        down_pt['cos_illumination'][down_pt['cos_illumination'] < 0] = 0
+        illumination_mask = down_pt['cos_illumination'] < 0
+        illumination_mask = illumination_mask.compute()
+        down_pt['cos_illumination'][illumination_mask] = 0
 
         # Binary shadow masks.
         horizon = horizon_da.sel(x=row.x, y=row.y, azimuth=np.rad2deg(ds_solar.azimuth), method='nearest')
