### 2023-06-13 bf5c520
update era5T realtime compute last date by simply removing embargo period from NOW (5days). Parsing the error msg was failing. This ECWMF thread covers the topic - watch this space to see if a simple api call to get last timestamp is created. https://confluence.ecmwf.int/pages/viewpage.action?pageId=277352608&focusedCommentId=278530169
