### 2023-06-13 3da2102
added function to write git commit hash and package version inside netcdf down_pt_*.nc
