### 2023-06-13 8371129
if realtime is true and declared then the end date is automatically updated in topoclass to make sure all months are present (in the case that a routine has not been run for over a month - for example. This is because the main realtime routine just updates the current month assuming that all previous months are present as it is intended to be run daily or so.
