### 2023-06-13 ed2866c
new ncol arg to construct_HX method so that variable can be specified more flexiblly. Plot routines refined.
