### 2023-06-13 5ee4f3d
column indexes of fsm output file updated as we no longer get an hour column
