### 2023-06-13 5af8c8d
simplified environment to the main packages
