### 2023-06-13 aa26b6a
added gdal, and pyModis in environment.yml file
