+++ environment.yml
@@ -7,6 +7,7 @@ dependencies:
   - dask==2023.4.1
   - dask-core=2023.4.1
   - geotiff=1.7.1
+  - gdal=3.6.4
   - h5netcdf=1.1.0
   - h5py=3.8.0
   - hdf4=4.2.15
@@ -37,3 +38,4 @@ dependencies:
       - scipy==1.10.1
       - seaborn==0.12.2
       - topocalc==0.5.0
+      - pyModis==2.3.0
