### 2023-06-13 8aadd96
added gitpython to environment.yml
