+++ environment.yml
@@ -39,3 +39,4 @@ dependencies:
       - seaborn==0.12.2
       - topocalc==0.5.0
       - pyModis==2.3.0
+      - gitpython==3.2.31
