### 2023-06-13 64c139f
debug writing version and git hash to dataset and netcdf
