### 2023-06-13 39f7030
added version and git commit to all types of netcdf exported from TopoPyscale
