### 2023-06-14 e0d3313
moved multithreading and multicore functions to topo_utils.py for general access across toolbox
