### 2023-06-14 91c8528
started function to parallelize FSM simulations
