### 2023-06-14 8d7d0ec
added conversion of FSM txt outputs to compressed netcdf, with function to process multipli files in parallel
