+++ TopoPyScale/topo_sim.py
@@ -20,6 +20,7 @@ from datetime import datetime
 from matplotlib.backends.backend_pdf import PdfPages
 import datetime as dt
 from TopoPyScale import topo_utils as tu
+from TopoPyScale import topo_export as te
 
 def fsm_nlst(nconfig, metfile, nave):
     """
@@ -112,6 +113,53 @@ def txt2ds(fname):
 
     return ds
 
+
+def to_netcdf(fname_fsm_sim, complevel=9):
+    ver_dict = tu.get_versionning()
+
+    ds = txt2ds(fname_fsm_sim)
+    ds.albedo.attrs = {'units':'ratio', 'standard_name':'albedo', 'long_name':'Surface Effective Albedo', '_FillValue': -9999999.0}
+    ds.runoff.attrs = {'units':'kg m-2', 'standard_name':'runoff', 'long_name':'Cumulated runoff from snow', '_FillValue': -9999999.0}
+    ds.snd.attrs = {'units':'m', 'standard_name':'snd', 'long_name':'Average snow depth', '_FillValue': -9999999.0}
+    ds.swe.attrs = {'units':'kg m-2', 'standard_name':'swe', 'long_name':'Average snow water equivalent', '_FillValue': -9999999.0}
+    ds.t_surface.attrs = {'units':'°C', 'standard_name':'t_surface', 'long_name':'Average surface temperature', '_FillValue': -9999999.0}
+    ds.t_surface.attrs = {'units':'°C', 'standard_name':'t_soil', 'long_name':'Average soil temperature at 20 cm depth', '_FillValue': -9999999.0}
+    ds.attrs = {'title':'FSM simulation outputs',
+                'source': 'Data downscaled with TopoPyScale and simulated with FSM',
+                'package_version':ver_dict.get('package_version'),
+                'git_commit': ver_dict.get('git_commit'),
+                'date_created':dt.datetime.now().strftime('%Y/%m/%d %H:%M:%S')}
+    fout = f"{fname_fsm_sim[:-4]}.nc"
+    te.to_netcdf(ds, fout, complevel=complevel)
+    print(f"File {fout} saved")
+
+
+def to_netcdf_parallel(fsm_sims='./fsm_sims/sim_FSM*.txt',
+                       n_core=6,
+                       complevel=9,
+                       delete_txt_files=False):
+    print('---> Run FSM simulation in parallel')
+    # 1. create all nlsit_files
+    if isinstance(fsm_sims, str):
+        flist = glob.glob(fsm_sims)
+    elif isinstance(fsm_sims, list):
+        flist = fsm_sims
+    else:
+        print('ERROR: fsm_input must either be a list of file path or a string of file pattern')
+        return
+
+    (ds, fname='output.nc', variables=None, complevel=9)
+    fun_param = zip(flist, [complevel]*len(flist))
+    tu.multicore_pooling(to_netcdf, fun_param, n_core)
+    print('---> All FSM simulations stored as netcdf')
+
+    if delete_txt_files:
+        print('---> Removing FSM simulation .txt file')
+        for file in flist:
+            os.remove(file)
+
+
+
 def to_dataset(fname_pattern='sim_FSM_pt*.txt', fsm_path = "./fsm_sims/"):
     '''
     Function to read FSM outputs of one simulation into a single dataset.
@@ -169,6 +217,19 @@ def fsm_sim_parallel(fsm_input='outputs/FSM_pt*.txt',
                      n_core=6,
                      n_thread=100,
                      delete_nlst_files=True):
+    '''
+    Function to run parallelised simulations of FSM
+
+    Args:
+        fsm_input (str or list): file pattern or list of file input to FSM. Note that must be in short relative path!. Default = 'outputs/FSM_pt*.txt'
[…]
