### 2023-06-14 938c50f
added package url to netcdf metadata
