### 2023-06-15 c0fb49f
update to realtime routine - updates sim end date to latest data date
