### 2023-06-15 b5cf36e
updated routine for obtaining latest date that era5T is available
