+++ JOSS/paper.md
@@ -46,7 +46,7 @@ Global climate reanalyses and projections are available worldwide for the past a
 
 ![Workflow of TopoPyScale processing pipeline.\label{fig:figure2}](figure2.png){ width=50% }
 
-`TopoPyScale` consists of a set of tools to be run in a processing pipeline following the diagram of \autoref{fig:figure2}. First, it takes a Digital Elevation Model (DEM), and climate data as inputs. The climate data must include air temperature, pressure, specific humidity, and meridional and zonal wind components from atmospheric pressure levels ranging from mean sea level to above the highest elevation of the DEM, as well as incoming shortwave radiation, incoming longwave radiation, and precipitation from the surface level. The next step is to compute terrain morphometrics such as slope, aspect, and sky view factor. Then, downscaling can be run in two modes, 1) *point*, or 2) *TopoSUB* downscaling. *Point* downscaling is used for a list of specific points (*e.g.* weather station locations), for which meteorological variables are downscaled only at the coordinates of given points. The *TopoSUB* approach [@fiddesTopoSUBToolEfficient2012] is instead executed for semi-distributed spatial downscaling (\autoref{fig:figure2}). The DEM is clustered based on morphometrics. *Toposcale*, the downscaling routine, is then run only for the centroid of each cluster or the list of given points. This method efficiently abstracts the DEM into a limited number of representative points as opposed to all the DEM grid cells, with potentially several orders of magnitude of computational effort saved. `Toposcale` uses 3D spatial interpolation and geometrical corrections to downscale the meteorological variables (see Table 1). The results can then be exported in a number of formats that are able to force specialized land surface models for snow, permafrost or hydrology via readily extendable plugins. Currently, this includes the models Cryogrid [@gmd-16-2607-2023], Crocus [@gmd-5-773-2012], SNOWPACK [@BARTELT2002123], Snowmodel [@ADistributedSnowEvolutionModelingSystemSnowModel], Geotop [@gmd-7-2831-2014], and the data assimilation toolkit MuSA [@gmd-15-9127-2022]. Finally, 1D results can be mapped back to the full DEM in order to generate spatially complete results. `TopoPyScale` also includes a toolbox to perform snow simulations with [FSM](https://github.com/RichardEssery/FSM) [@gmd-8-3867-2015], and data assimilation of fractional snow-covered area [@hess-23-4717-2019].
[…]
