### 2023-06-15 #5059 @ArcticSnow
Thank you @hugoledoux

- right it sounds odd. I modified it to this `The results can then be exported in a number of formats that can be used as forcing for specialized land surface models for snow,...`
- I create a new release tag [v0.2.3](https://github.com/ArcticSnow/TopoPyScale/releases/tag/v0.2.3). It is also pushed on [Pypi](https://pypi.org/project/topopyscale/)
- archived on [Zenodo](https://zenodo.org/record/8043606) at this  DOI:10.5281/zenodo.8043606

-[x] Make sure that the license listed for the archive is the same as the software license. License is: MIT License

Please let me know if anything else is needed or missing.
Thank you
