### 2023-06-15 f61bdca
correct JOSS paper.md typo line 13
