### 2023-06-18 46c1495
moved steep_snow_reduce_all() to meteo_utils.py
