### 2023-06-21 9b36b16
updated routine for computing snow climatology
