### 2023-06-21 3857f2f
Implemented a more robust indexing of fsm files from last column (ie -1,-2,-3 etc) as compilations of fsm seem to sometimes produce hour columns (eg AWS) and sometimes not (eg WSL Hyperion) - therefore indexing from first column indexes different variables in these two cases.
