### 2023-06-21 bcccdd1
removed fsm file column number ref in snow climatology routine
