### 2023-06-22 #5059 @kthyng
The paper looks good except for capitalization in the references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, Kronenberg et al seems to have a proper noun in the title that is not capitalized.
