+++ JOSS/paper.bib
@@ -23,7 +23,7 @@ DOI = {10.5194/hess-23-4717-2019}
 
 @Article{tc-2021-380,
 AUTHOR = {Kronenberg, M. and van Pelt, W. and Machguth, H. and Fiddes, J. and Hoelzle, M. and Pertziger, F.},
-TITLE = {Long-term firn and mass balance modelling for Abramov glacier, Pamir Alay},
+TITLE = {Long-term firn and mass balance modelling for {A}bramov glacier, {P}amir {A}lay},
 JOURNAL = {The Cryosphere Discussions},
 VOLUME = {2022},
 YEAR = {2022},
@@ -34,7 +34,7 @@ DOI = {10.5194/tc-2021-380}
 
 @Article{esurf-2022-39,
 AUTHOR = {Renette, C. and Aalstad, K. and Aga, J. and Zweigel, R. B. and Etzelm\"uller, B. and Lille{\o}ren, K. S. and Isaksen, K. and Westermann, S.},
-TITLE = {Simulating the effect of subsurface drainage on the thermal regime and ground ice in blocky terrain, Norway},
+TITLE = {Simulating the effect of subsurface drainage on the thermal regime and ground ice in blocky terrain, {N}orway},
 JOURNAL = {Earth Surface Dynamics Discussions},
 VOLUME = {2022},
 YEAR = {2022},
@@ -45,7 +45,7 @@ DOI = {10.5194/esurf-2022-39}
 
 @Article{gmd-15-9127-2022,
 AUTHOR = {Alonso-Gonz\'alez, E. and Aalstad, K. and Baba, M. W. and Revuelto, J. and L\'opez-Moreno, J. I. and Fiddes, J. and Essery, R. and Gascoin, S.},
-TITLE = {The Multiple Snow Data Assimilation System (MuSA v1.0)},
+TITLE = {The {M}ultiple {S}now {D}ata {A}ssimilation {S}ystem (MuSA v1.0)},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {15},
 YEAR = {2022},
@@ -145,7 +145,7 @@ year = {2019}
 
 @Article{gmd-5-773-2012,
 AUTHOR = {Vionnet, V. and Brun, E. and Morin, S. and Boone, A. and Faroux, S. and Le Moigne, P. and Martin, E. and Willemet, J.-M.},
-TITLE = {The detailed snowpack scheme Crocus and its implementation in SURFEX v7.2},
+TITLE = {The detailed snowpack scheme {C}rocus and its implementation in {SURFEX} v7.2},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {5},
 YEAR = {2012},
@@ -156,7 +156,7 @@ DOI = {10.5194/gmd-5-773-2012}
 }
 
 @article{BARTELT2002123,
-title = {A physical SNOWPACK model for the Swiss avalanche warning: Part I: numerical model},
+title = {A physical SNOWPACK model for the {S}wiss avalanche warning: Part I: numerical model},
 journal = {Cold Regions Science and Technology},
 volume = {35},
 number = {3},
@@ -265,7 +265,7 @@ keywords = {Snow, Heat and mass transfer, Snow deformation, Phase change, Finite
 
 @article {ADistributedSnowEvolutionModelingSystemSnowModel,
       author = "Glen E.  Liston and Kelly  Elder",
-      title = "A Distributed Snow-Evolution Modeling System (SnowModel)",
+      title = "A Distributed Snow-Evolution Modeling System {{(SnowModel)}}",
       journal = "Journal of Hydrometeorology",
       year = "2006",
       publisher = "American Meteorological Society",
[…]
