+++ JOSS/paper.bib
@@ -45,7 +45,7 @@ DOI = {10.5194/esurf-2022-39}
 
 @Article{gmd-15-9127-2022,
 AUTHOR = {Alonso-Gonz\'alez, E. and Aalstad, K. and Baba, M. W. and Revuelto, J. and L\'opez-Moreno, J. I. and Fiddes, J. and Essery, R. and Gascoin, S.},
-TITLE = {The {M}ultiple {S}now {D}ata {A}ssimilation {S}ystem (MuSA v1.0)},
+TITLE = {The {M}ultiple {S}now {D}ata {A}ssimilation {S}ystem ({{MuSA}} v1.0)},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {15},
 YEAR = {2022},
@@ -58,7 +58,7 @@ DOI = {10.5194/gmd-15-9127-2022}
 
 @Article{gmd-8-3867-2015,
 AUTHOR = {Essery, R.},
-TITLE = {A factorial snowpack model (FSM 1.0)},
+TITLE = {A factorial snowpack model ({{FSM}} 1.0)},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {8},
 YEAR = {2015},
@@ -156,7 +156,7 @@ DOI = {10.5194/gmd-5-773-2012}
 }
 
 @article{BARTELT2002123,
-title = {A physical SNOWPACK model for the {S}wiss avalanche warning: Part I: numerical model},
+title = {A physical {{SNOWPACK}} model for the {S}wiss avalanche warning: Part I: numerical model},
 journal = {Cold Regions Science and Technology},
 volume = {35},
 number = {3},
