### 2023-06-24 #5059 @ArcticSnow
Thank you @kthyng for finalizing the publication of the manuscript.

Thank you also to @hugoledoux , @arbennett  and @dvalters for a constructive reviewing process. I enjoyed the emphasis on publishing a user-focused paper/software package. It forced us to put in place a consistent platform for using and developing this package

I will gladly take part to the JOSS adventure and will sign up for joining as a reviewer.

Thank you from me and my co-authors!
