### 2025-07-15 #5059 @paulrougieux
@ArcticSnow your package uses Xarray as well. Would you like to review my paper [Cobwood: Enhancing Forest Economics Model Reusability Through Labeled Panel Data Structures](https://github.com/openjournals/joss-reviews/issues/8587#issuecomment-3071248774) ?
